% Auto-generated: do not edit by hand
\name{accordion}

\alias{accordion}

\title{Accordion component}

\description{
Accordion is a wrapper for the Carbon Accordion component.
}

\usage{
accordion(children=NULL, id=NULL, align=NULL, className=NULL,
disabled=NULL, isFlush=NULL, loading_state=NULL,
ordered=NULL, size=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Character. align}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{isFlush}{Logical. isFlush}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{ordered}{Logical | numeric | character | named list | unnamed list. ordered}

\item{size}{Character. size}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

