% Auto-generated: do not edit by hand
\name{accordionItem}

\alias{accordionItem}

\title{AccordionItem component}

\description{
AccordionItem is a wrapper for the Carbon AccordionItem component.
}

\usage{
accordionItem(children=NULL, id=NULL, className=NULL, disabled=NULL,
loading_state=NULL, onClick=NULL, onHeadingClick=NULL,
open=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, renderExpando=NULL,
renderToggle=NULL, style=NULL, title=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onHeadingClick}{Logical | numeric | character | named list | unnamed list. onHeadingClick}

\item{open}{Logical. open}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{renderExpando}{Logical | numeric | character | named list | unnamed list. renderExpando}

\item{renderToggle}{Logical | numeric | character | named list | unnamed list. renderToggle}

\item{style}{Named list. style}

\item{title}{A list of or a singular dash component, string or number. title}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

