% Auto-generated: do not edit by hand
\name{accordionSkeleton}

\alias{accordionSkeleton}

\title{AccordionSkeleton component}

\description{
AccordionSkeleton is a wrapper for the Carbon AccordionSkeleton component.
}

\usage{
accordionSkeleton(children=NULL, id=NULL, align=NULL, className=NULL,
count=NULL, isFlush=NULL, loading_state=NULL, open=NULL,
ordered=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{className}{Character. className}

\item{count}{Logical | numeric | character | named list | unnamed list. count}

\item{isFlush}{Logical | numeric | character | named list | unnamed list. isFlush}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{open}{Logical. open}

\item{ordered}{Logical. ordered}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

