% Auto-generated: do not edit by hand
\name{aILabel}

\alias{aILabel}

\title{AILabel component}

\description{
AILabel is a wrapper for the Carbon AILabel component.
}

\usage{
aILabel(children=NULL, id=NULL, AILabelContent=NULL, aiText=NULL,
aiTextLabel=NULL, align=NULL, alignmentAxisOffset=NULL,
as_=NULL, autoAlign=NULL, autoAlignBoundary=NULL,
backgroundToken=NULL, border=NULL, caret=NULL,
className=NULL, defaultOpen=NULL, dropShadow=NULL,
highContrast=NULL, isTabTip=NULL, kind=NULL,
loading_state=NULL, onRequestClose=NULL, onRevertClick=NULL,
revertActive=NULL, revertLabel=NULL, size=NULL,
slugLabel=NULL, style=NULL, textLabel=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{AILabelContent}{Logical | numeric | character | named list | unnamed list. AILabelContent}

\item{aiText}{Logical | numeric | character | named list | unnamed list. aiText}

\item{aiTextLabel}{Logical | numeric | character | named list | unnamed list. aiTextLabel}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{alignmentAxisOffset}{Logical | numeric | character | named list | unnamed list. alignmentAxisOffset}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{autoAlignBoundary}{Logical | numeric | character | named list | unnamed list. autoAlignBoundary}

\item{backgroundToken}{Logical | numeric | character | named list | unnamed list. backgroundToken}

\item{border}{Logical | numeric | character | named list | unnamed list. border}

\item{caret}{Logical | numeric | character | named list | unnamed list. caret}

\item{className}{Character. className}

\item{defaultOpen}{Logical | numeric | character | named list | unnamed list. defaultOpen}

\item{dropShadow}{Logical | numeric | character | named list | unnamed list. dropShadow}

\item{highContrast}{Logical | numeric | character | named list | unnamed list. highContrast}

\item{isTabTip}{Logical | numeric | character | named list | unnamed list. isTabTip}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onRequestClose}{Logical | numeric | character | named list | unnamed list. onRequestClose}

\item{onRevertClick}{Logical | numeric | character | named list | unnamed list. onRevertClick}

\item{revertActive}{Logical | numeric | character | named list | unnamed list. revertActive}

\item{revertLabel}{Logical | numeric | character | named list | unnamed list. revertLabel}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slugLabel}{Logical | numeric | character | named list | unnamed list. slugLabel}

\item{style}{Named list. style}

\item{textLabel}{Logical | numeric | character | named list | unnamed list. textLabel}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

