% Auto-generated: do not edit by hand
\name{aISkeletonText}

\alias{aISkeletonText}

\title{AISkeletonText component}

\description{
AISkeletonText is a wrapper for the Carbon AISkeletonText component.
}

\usage{
aISkeletonText(children=NULL, id=NULL, className=NULL, heading=NULL,
lineCount=NULL, loading_state=NULL, paragraph=NULL,
style=NULL, width=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{heading}{Logical | numeric | character | named list | unnamed list. heading}

\item{lineCount}{Logical | numeric | character | named list | unnamed list. lineCount}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{paragraph}{Logical | numeric | character | named list | unnamed list. paragraph}

\item{style}{Named list. style}

\item{width}{Logical | numeric | character | named list | unnamed list. width}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

