% Auto-generated: do not edit by hand
\name{breadcrumbItem}

\alias{breadcrumbItem}

\title{BreadcrumbItem component}

\description{
BreadcrumbItem is a wrapper for the Carbon BreadcrumbItem component.
}

\usage{
breadcrumbItem(children=NULL, id=NULL, className=NULL, href=NULL,
isCurrentPage=NULL, loading_state=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{href}{Character. href}

\item{isCurrentPage}{Logical. isCurrentPage}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

