% Auto-generated: do not edit by hand
\name{breadcrumbSkeleton}

\alias{breadcrumbSkeleton}

\title{BreadcrumbSkeleton component}

\description{
BreadcrumbSkeleton is a wrapper for the Carbon BreadcrumbSkeleton component.
}

\usage{
breadcrumbSkeleton(children=NULL, id=NULL, className=NULL, items=NULL,
loading_state=NULL, noTrailingSlash=NULL, size=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{items}{Logical | numeric | character | named list | unnamed list. items}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{noTrailingSlash}{Logical | numeric | character | named list | unnamed list. noTrailingSlash}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

