% Auto-generated: do not edit by hand
\name{button}

\alias{button}

\title{Button component}

\description{
Button is a wrapper for the Carbon Button component.
}

\usage{
button(children=NULL, id=NULL, as_=NULL, autoAlign=NULL,
className=NULL, dangerDescription=NULL, disabled=NULL,
hasIconOnly=NULL, href=NULL, iconDescription=NULL,
isExpressive=NULL, isSelected=NULL, kind=NULL,
loading_state=NULL, n_clicks=NULL, onBlur=NULL,
onClick=NULL, onFocus=NULL, onMouseEnter=NULL,
onMouseLeave=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, rel=NULL, renderIcon=NULL, role=NULL,
size=NULL, style=NULL, tabIndex=NULL, target=NULL,
tooltipAlignment=NULL, tooltipDropShadow=NULL,
tooltipHighContrast=NULL, tooltipPosition=NULL, type=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{className}{Character. className}

\item{dangerDescription}{Logical | numeric | character | named list | unnamed list. dangerDescription}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{hasIconOnly}{Logical | numeric | character | named list | unnamed list. hasIconOnly}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{isExpressive}{Logical | numeric | character | named list | unnamed list. isExpressive}

\item{isSelected}{Logical | numeric | character | named list | unnamed list. isSelected}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_clicks}{Numeric. n_clicks}

\item{onBlur}{Logical | numeric | character | named list | unnamed list. onBlur}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onFocus}{Logical | numeric | character | named list | unnamed list. onFocus}

\item{onMouseEnter}{Logical | numeric | character | named list | unnamed list. onMouseEnter}

\item{onMouseLeave}{Logical | numeric | character | named list | unnamed list. onMouseLeave}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{rel}{Logical | numeric | character | named list | unnamed list. rel}

\item{renderIcon}{Logical | numeric | character | named list | unnamed list. renderIcon}

\item{role}{Logical | numeric | character | named list | unnamed list. role}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}

\item{target}{Logical | numeric | character | named list | unnamed list. target}

\item{tooltipAlignment}{Logical | numeric | character | named list | unnamed list. tooltipAlignment}

\item{tooltipDropShadow}{Logical | numeric | character | named list | unnamed list. tooltipDropShadow}

\item{tooltipHighContrast}{Logical | numeric | character | named list | unnamed list. tooltipHighContrast}

\item{tooltipPosition}{Logical | numeric | character | named list | unnamed list. tooltipPosition}

\item{type}{Logical | numeric | character | named list | unnamed list. type}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

