% Auto-generated: do not edit by hand
\name{buttonSet}

\alias{buttonSet}

\title{ButtonSet component}

\description{
ButtonSet is a wrapper for the Carbon ButtonSet component.
}

\usage{
buttonSet(children=NULL, id=NULL, className=NULL, fluid=NULL,
loading_state=NULL, stacked=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{fluid}{Logical | numeric | character | named list | unnamed list. fluid}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{stacked}{Logical | numeric | character | named list | unnamed list. stacked}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

