% Auto-generated: do not edit by hand
\name{buttonSizes}

\alias{buttonSizes}

\title{ButtonSizes component}

\description{
ButtonSizes is a wrapper for the Carbon ButtonSizes component.
}

\usage{
buttonSizes(children=NULL, id=NULL, className=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The content of the component.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Specify a custom className to be applied to the component.}

\item{style}{Named list. Inline styles}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

