% Auto-generated: do not edit by hand
\name{callout}

\alias{callout}

\title{Callout component}

\description{
Callout is a wrapper for the Carbon Callout component.
}

\usage{
callout(children=NULL, id=NULL, actionButtonLabel=NULL,
className=NULL, kind=NULL, loading_state=NULL,
lowContrast=NULL, onActionButtonClick=NULL,
statusIconDescription=NULL, style=NULL, subtitle=NULL,
title=NULL, titleId=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{actionButtonLabel}{Logical | numeric | character | named list | unnamed list. actionButtonLabel}

\item{className}{Character. className}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{lowContrast}{Logical | numeric | character | named list | unnamed list. lowContrast}

\item{onActionButtonClick}{Logical | numeric | character | named list | unnamed list. onActionButtonClick}

\item{statusIconDescription}{Logical | numeric | character | named list | unnamed list. statusIconDescription}

\item{style}{Named list. style}

\item{subtitle}{Logical | numeric | character | named list | unnamed list. subtitle}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{titleId}{Logical | numeric | character | named list | unnamed list. titleId}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

