% Auto-generated: do not edit by hand
\name{chatButton}

\alias{chatButton}

\title{ChatButton component}

\description{
ChatButton is a wrapper for the Carbon ChatButton component.
}

\usage{
chatButton(children=NULL, id=NULL, className=NULL, disabled=NULL,
isQuickAction=NULL, isSelected=NULL, kind=NULL,
loading_state=NULL, renderIcon=NULL, size=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{isQuickAction}{Logical | numeric | character | named list | unnamed list. isQuickAction}

\item{isSelected}{Logical | numeric | character | named list | unnamed list. isSelected}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

