% Auto-generated: do not edit by hand
\name{checkbox}

\alias{checkbox}

\title{Checkbox component}

\description{
Checkbox is a wrapper for the Carbon Checkbox component.
}

\usage{
checkbox(children=NULL, id=NULL, checked=NULL, className=NULL,
debounce=NULL, decorator=NULL, defaultChecked=NULL,
disabled=NULL, helperText=NULL, hideLabel=NULL,
indeterminate=NULL, invalid=NULL, invalidText=NULL,
label=NULL, labelText=NULL, loading_state=NULL, n_blur=NULL,
n_submit=NULL, onChange=NULL, persisted_props=NULL,
persistence=NULL, persistence_type=NULL, readOnly=NULL,
slug=NULL, style=NULL, title=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{checked}{Logical. checked}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{defaultChecked}{Logical | numeric | character | named list | unnamed list. defaultChecked}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{indeterminate}{Logical | numeric | character | named list | unnamed list. indeterminate}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{label}{Character. label}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

