% Auto-generated: do not edit by hand
\name{checkboxGroup}

\alias{checkboxGroup}

\title{CheckboxGroup component}

\description{
CheckboxGroup is a wrapper for the Carbon CheckboxGroup component.
}

\usage{
checkboxGroup(children=NULL, id=NULL, className=NULL, decorator=NULL,
helperText=NULL, invalid=NULL, invalidText=NULL,
legendId=NULL, legendText=NULL, loading_state=NULL,
orientation=NULL, readOnly=NULL, slug=NULL, style=NULL,
warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{legendId}{Logical | numeric | character | named list | unnamed list. legendId}

\item{legendText}{Logical | numeric | character | named list | unnamed list. legendText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{orientation}{Logical | numeric | character | named list | unnamed list. orientation}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

