% Auto-generated: do not edit by hand
\name{clickableTile}

\alias{clickableTile}

\title{ClickableTile component}

\description{
ClickableTile is a wrapper for the Carbon ClickableTile component.
}

\usage{
clickableTile(children=NULL, id=NULL, className=NULL, clicked=NULL,
decorator=NULL, disabled=NULL, hasRoundedCorners=NULL,
href=NULL, light=NULL, loading_state=NULL, onClick=NULL,
onKeyDown=NULL, rel=NULL, renderIcon=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{clicked}{Logical | numeric | character | named list | unnamed list. clicked}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{hasRoundedCorners}{Logical | numeric | character | named list | unnamed list. hasRoundedCorners}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{rel}{Logical | numeric | character | named list | unnamed list. rel}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

