% Auto-generated: do not edit by hand
\name{codeSnippet}

\alias{codeSnippet}

\title{CodeSnippet component}

\description{
CodeSnippet is a wrapper for the Carbon CodeSnippet component.
}

\usage{
codeSnippet(children=NULL, id=NULL, align=NULL, ariaLabel=NULL,
autoAlign=NULL, className=NULL, copyButtonDescription=NULL,
copyText=NULL, disabled=NULL, feedback=NULL,
feedbackTimeout=NULL, hideCopyButton=NULL, light=NULL,
loading_state=NULL, maxCollapsedNumberOfRows=NULL,
maxExpandedNumberOfRows=NULL, minCollapsedNumberOfRows=NULL,
minExpandedNumberOfRows=NULL, onClick=NULL,
showLessText=NULL, showMoreText=NULL, style=NULL, type=NULL,
wrapText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{className}{Character. className}

\item{copyButtonDescription}{Logical | numeric | character | named list | unnamed list. copyButtonDescription}

\item{copyText}{Logical | numeric | character | named list | unnamed list. copyText}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{feedback}{Logical | numeric | character | named list | unnamed list. feedback}

\item{feedbackTimeout}{Logical | numeric | character | named list | unnamed list. feedbackTimeout}

\item{hideCopyButton}{Logical | numeric | character | named list | unnamed list. hideCopyButton}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{maxCollapsedNumberOfRows}{Logical | numeric | character | named list | unnamed list. maxCollapsedNumberOfRows}

\item{maxExpandedNumberOfRows}{Logical | numeric | character | named list | unnamed list. maxExpandedNumberOfRows}

\item{minCollapsedNumberOfRows}{Logical | numeric | character | named list | unnamed list. minCollapsedNumberOfRows}

\item{minExpandedNumberOfRows}{Logical | numeric | character | named list | unnamed list. minExpandedNumberOfRows}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{showLessText}{Logical | numeric | character | named list | unnamed list. showLessText}

\item{showMoreText}{Logical | numeric | character | named list | unnamed list. showMoreText}

\item{style}{Named list. style}

\item{type}{Logical | numeric | character | named list | unnamed list. type}

\item{wrapText}{Logical | numeric | character | named list | unnamed list. wrapText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

