% Auto-generated: do not edit by hand
\name{column}

\alias{column}

\title{Column component}

\description{
Column is a wrapper for the Carbon Column component.
}

\usage{
column(children=NULL, id=NULL, as_=NULL, className=NULL, lg=NULL,
loading_state=NULL, max=NULL, md=NULL, sm=NULL, span=NULL,
style=NULL, xlg=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{lg}{Logical | numeric | character | named list | unnamed list. lg}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{max}{Logical | numeric | character | named list | unnamed list. max}

\item{md}{Logical | numeric | character | named list | unnamed list. md}

\item{sm}{Logical | numeric | character | named list | unnamed list. sm}

\item{span}{Logical | numeric | character | named list | unnamed list. span}

\item{style}{Named list. style}

\item{xlg}{Logical | numeric | character | named list | unnamed list. xlg}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

