% Auto-generated: do not edit by hand
\name{comboBox}

\alias{comboBox}

\title{ComboBox component}

\description{
ComboBox is a wrapper for the Carbon ComboBox component.
}

\usage{
comboBox(children=NULL, id=NULL, allowCustomValue=NULL,
ariaLabel=NULL, autoAlign=NULL, className=NULL,
decorator=NULL, direction=NULL, disabled=NULL,
downshiftActions=NULL, downshiftProps=NULL, helperText=NULL,
initialSelectedItem=NULL, inputProps=NULL, invalid=NULL,
invalidText=NULL, itemToElement=NULL, itemToString=NULL,
items=NULL, light=NULL, loading_state=NULL, onChange=NULL,
onInputChange=NULL, onToggleClick=NULL, placeholder=NULL,
readOnly=NULL, selectedItem=NULL, shouldFilterItem=NULL,
size=NULL, slug=NULL, style=NULL, titleText=NULL,
translateWithId=NULL, typeahead=NULL, warn=NULL,
warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{allowCustomValue}{Logical | numeric | character | named list | unnamed list. allowCustomValue}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{direction}{Logical | numeric | character | named list | unnamed list. direction}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{downshiftActions}{Logical | numeric | character | named list | unnamed list. downshiftActions}

\item{downshiftProps}{Logical | numeric | character | named list | unnamed list. downshiftProps}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{initialSelectedItem}{Logical | numeric | character | named list | unnamed list. initialSelectedItem}

\item{inputProps}{Logical | numeric | character | named list | unnamed list. inputProps}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{itemToElement}{Logical | numeric | character | named list | unnamed list. itemToElement}

\item{itemToString}{Logical | numeric | character | named list | unnamed list. itemToString}

\item{items}{Logical | numeric | character | named list | unnamed list. items}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onInputChange}{Logical | numeric | character | named list | unnamed list. onInputChange}

\item{onToggleClick}{Logical | numeric | character | named list | unnamed list. onToggleClick}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{selectedItem}{Logical | numeric | character | named list | unnamed list. selectedItem}

\item{shouldFilterItem}{Logical | numeric | character | named list | unnamed list. shouldFilterItem}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{titleText}{Logical | numeric | character | named list | unnamed list. titleText}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}

\item{typeahead}{Logical | numeric | character | named list | unnamed list. typeahead}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

