% Auto-generated: do not edit by hand
\name{comboButton}

\alias{comboButton}

\title{ComboButton component}

\description{
ComboButton is a wrapper for the Carbon ComboButton component.
}

\usage{
comboButton(children=NULL, id=NULL, className=NULL, disabled=NULL,
label=NULL, loading_state=NULL, menuAlignment=NULL,
onClick=NULL, size=NULL, style=NULL, tooltipAlignment=NULL,
translateWithId=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{menuAlignment}{Logical | numeric | character | named list | unnamed list. menuAlignment}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{tooltipAlignment}{Logical | numeric | character | named list | unnamed list. tooltipAlignment}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

