% Auto-generated: do not edit by hand
\name{composedModal}

\alias{composedModal}

\title{ComposedModal component}

\description{
ComposedModal is a wrapper for the Carbon ComposedModal component.
}

\usage{
composedModal(children=NULL, id=NULL, className=NULL,
containerClassName=NULL, current=NULL, danger=NULL,
decorator=NULL, isFullWidth=NULL, launcherButtonRef=NULL,
loading_state=NULL, onClose=NULL, onKeyDown=NULL, open=NULL,
preventCloseOnClickOutside=NULL, selectorPrimaryFocus=NULL,
selectorsFloatingMenus=NULL, size=NULL, slug=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{containerClassName}{Logical | numeric | character | named list | unnamed list. containerClassName}

\item{current}{Logical | numeric | character | named list | unnamed list. current}

\item{danger}{Logical | numeric | character | named list | unnamed list. danger}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{isFullWidth}{Logical | numeric | character | named list | unnamed list. isFullWidth}

\item{launcherButtonRef}{Logical | numeric | character | named list | unnamed list. launcherButtonRef}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClose}{Logical | numeric | character | named list | unnamed list. onClose}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{open}{Logical. open}

\item{preventCloseOnClickOutside}{Logical | numeric | character | named list | unnamed list. preventCloseOnClickOutside}

\item{selectorPrimaryFocus}{Logical | numeric | character | named list | unnamed list. selectorPrimaryFocus}

\item{selectorsFloatingMenus}{Logical | numeric | character | named list | unnamed list. selectorsFloatingMenus}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

