% Auto-generated: do not edit by hand
\name{containedListItem}

\alias{containedListItem}

\title{ContainedListItem component}

\description{
ContainedListItem is a wrapper for the Carbon ContainedListItem component.
}

\usage{
containedListItem(children=NULL, id=NULL, action=NULL, className=NULL,
disabled=NULL, loading_state=NULL, onClick=NULL,
renderIcon=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{action}{Logical | numeric | character | named list | unnamed list. action}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

