% Auto-generated: do not edit by hand
\name{contentSwitcher}

\alias{contentSwitcher}

\title{ContentSwitcher component}

\description{
ContentSwitcher is a wrapper for the Carbon ContentSwitcher component.
}

\usage{
contentSwitcher(children=NULL, id=NULL, className=NULL, light=NULL,
loading_state=NULL, lowContrast=NULL, onChange=NULL,
selectedIndex=NULL, selectionMode=NULL, size=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{lowContrast}{Logical | numeric | character | named list | unnamed list. lowContrast}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{selectedIndex}{Numeric. selectedIndex}

\item{selectionMode}{Logical | numeric | character | named list | unnamed list. selectionMode}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

