% Auto-generated: do not edit by hand
\name{contextMenu}

\alias{contextMenu}

\title{ContextMenu component}

\description{
ContextMenu is a wrapper for the Carbon ContextMenu component.
}

\usage{
contextMenu(children=NULL, id=NULL, className=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The content of the component.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Specify a custom className to be applied to the component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

