% Auto-generated: do not edit by hand
\name{controlledPasswordInput}

\alias{controlledPasswordInput}

\title{ControlledPasswordInput component}

\description{
ControlledPasswordInput is a wrapper for the Carbon ControlledPasswordInput component.
}

\usage{
controlledPasswordInput(children=NULL, id=NULL, className=NULL, debounce=NULL,
defaultValue=NULL, disabled=NULL, helperText=NULL,
hideLabel=NULL, hidePasswordLabel=NULL, invalid=NULL,
invalidText=NULL, labelText=NULL, light=NULL,
loading_state=NULL, n_blur=NULL, n_submit=NULL,
onChange=NULL, onClick=NULL, persisted_props=NULL,
persistence=NULL, persistence_type=NULL, placeholder=NULL,
showPasswordLabel=NULL, size=NULL, style=NULL,
tooltipAlignment=NULL, tooltipPosition=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. defaultValue}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{hidePasswordLabel}{Logical | numeric | character | named list | unnamed list. hidePasswordLabel}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{showPasswordLabel}{Logical | numeric | character | named list | unnamed list. showPasswordLabel}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{tooltipAlignment}{Logical | numeric | character | named list | unnamed list. tooltipAlignment}

\item{tooltipPosition}{Logical | numeric | character | named list | unnamed list. tooltipPosition}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

