% Auto-generated: do not edit by hand
\name{copy}

\alias{copy}

\title{Copy component}

\description{
Copy is a wrapper for the Carbon Copy component.
}

\usage{
copy(children=NULL, id=NULL, align=NULL, autoAlign=NULL,
className=NULL, feedback=NULL, feedbackTimeout=NULL,
loading_state=NULL, onAnimationEnd=NULL, onClick=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{className}{Character. className}

\item{feedback}{Logical | numeric | character | named list | unnamed list. feedback}

\item{feedbackTimeout}{Logical | numeric | character | named list | unnamed list. feedbackTimeout}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onAnimationEnd}{Logical | numeric | character | named list | unnamed list. onAnimationEnd}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

