% Auto-generated: do not edit by hand
\name{dataTable}

\alias{dataTable}

\title{DataTable component}

\description{
DataTable is a wrapper for the Carbon DataTable component.
}

\usage{
dataTable(children=NULL, id=NULL, className=NULL, description=NULL,
experimentalAutoAlign=NULL, filterRows=NULL, header=NULL,
headers=NULL, isSortable=NULL, key=NULL, loading_state=NULL,
locale=NULL, overflowMenuOnHover=NULL, radio=NULL,
render=NULL, rows=NULL, size=NULL, sortRow=NULL,
stickyHeader=NULL, style=NULL, title=NULL,
translateWithId=NULL, useStaticWidth=NULL,
useZebraStyles=NULL, withExpansion=NULL, withSelection=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{description}{Character. description}

\item{experimentalAutoAlign}{Logical | numeric | character | named list | unnamed list. experimentalAutoAlign}

\item{filterRows}{Logical | numeric | character | named list | unnamed list. filterRows}

\item{header}{Logical | numeric | character | named list | unnamed list. header}

\item{headers}{Unnamed list. headers}

\item{isSortable}{Logical. isSortable}

\item{key}{Logical | numeric | character | named list | unnamed list. key}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{locale}{Logical | numeric | character | named list | unnamed list. locale}

\item{overflowMenuOnHover}{Logical | numeric | character | named list | unnamed list. overflowMenuOnHover}

\item{radio}{Logical | numeric | character | named list | unnamed list. radio}

\item{render}{Logical | numeric | character | named list | unnamed list. render}

\item{rows}{Unnamed list. rows}

\item{size}{Character. size}

\item{sortRow}{Logical | numeric | character | named list | unnamed list. sortRow}

\item{stickyHeader}{Logical | numeric | character | named list | unnamed list. stickyHeader}

\item{style}{Named list. style}

\item{title}{Character. title}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}

\item{useStaticWidth}{Logical | numeric | character | named list | unnamed list. useStaticWidth}

\item{useZebraStyles}{Logical. useZebraStyles}

\item{withExpansion}{Logical. withExpansion}

\item{withSelection}{Logical. withSelection}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

