% Auto-generated: do not edit by hand
\name{dataTableSkeleton}

\alias{dataTableSkeleton}

\title{DataTableSkeleton component}

\description{
DataTableSkeleton is a wrapper for the Carbon DataTableSkeleton component.
}

\usage{
dataTableSkeleton(children=NULL, id=NULL, className=NULL, columnCount=NULL,
headers=NULL, loading_state=NULL, rowCount=NULL,
showHeader=NULL, showToolbar=NULL, size=NULL, style=NULL,
zebra=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{columnCount}{Logical | numeric | character | named list | unnamed list. columnCount}

\item{headers}{Logical | numeric | character | named list | unnamed list. headers}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{rowCount}{Logical | numeric | character | named list | unnamed list. rowCount}

\item{showHeader}{Logical | numeric | character | named list | unnamed list. showHeader}

\item{showToolbar}{Logical | numeric | character | named list | unnamed list. showToolbar}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{zebra}{Logical | numeric | character | named list | unnamed list. zebra}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

