% Auto-generated: do not edit by hand
\name{datePicker}

\alias{datePicker}

\title{DatePicker component}

\description{
DatePicker is a wrapper for the Carbon DatePicker component.
}

\usage{
datePicker(children=NULL, id=NULL, allowInput=NULL, appendTo=NULL,
className=NULL, closeOnSelect=NULL, dateFormat=NULL,
datePickerType=NULL, debounce=NULL, disable=NULL,
enable=NULL, inline=NULL, invalid=NULL, invalidText=NULL,
light=NULL, loading_state=NULL, locale=NULL, maxDate=NULL,
minDate=NULL, n_blur=NULL, n_submit=NULL,
nextMonthAriaLabel=NULL, onChange=NULL, onClose=NULL,
onOpen=NULL, parseDate=NULL, persisted_props=NULL,
persistence=NULL, persistence_type=NULL,
prevMonthAriaLabel=NULL, readOnly=NULL, short=NULL,
style=NULL, value=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{allowInput}{Logical | numeric | character | named list | unnamed list. allowInput}

\item{appendTo}{Logical | numeric | character | named list | unnamed list. appendTo}

\item{className}{Character. className}

\item{closeOnSelect}{Logical | numeric | character | named list | unnamed list. closeOnSelect}

\item{dateFormat}{Logical | numeric | character | named list | unnamed list. dateFormat}

\item{datePickerType}{Character. datePickerType}

\item{debounce}{Logical | numeric. debounce}

\item{disable}{Logical | numeric | character | named list | unnamed list. disable}

\item{enable}{Logical | numeric | character | named list | unnamed list. enable}

\item{inline}{Logical | numeric | character | named list | unnamed list. inline}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{locale}{Logical | numeric | character | named list | unnamed list. locale}

\item{maxDate}{Logical | numeric | character | named list | unnamed list. maxDate}

\item{minDate}{Logical | numeric | character | named list | unnamed list. minDate}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{nextMonthAriaLabel}{Logical | numeric | character | named list | unnamed list. nextMonthAriaLabel}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClose}{Logical | numeric | character | named list | unnamed list. onClose}

\item{onOpen}{Logical | numeric | character | named list | unnamed list. onOpen}

\item{parseDate}{Logical | numeric | character | named list | unnamed list. parseDate}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{prevMonthAriaLabel}{Logical | numeric | character | named list | unnamed list. prevMonthAriaLabel}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{short}{Logical | numeric | character | named list | unnamed list. short}

\item{style}{Named list. style}

\item{value}{Logical | numeric | character | named list | unnamed list. value}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

