% Auto-generated: do not edit by hand
\name{datePickerSkeleton}

\alias{datePickerSkeleton}

\title{DatePickerSkeleton component}

\description{
DatePickerSkeleton is a wrapper for the Carbon DatePickerSkeleton component.
}

\usage{
datePickerSkeleton(children=NULL, id=NULL, className=NULL, hideLabel=NULL,
loading_state=NULL, range=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{range}{Logical | numeric | character | named list | unnamed list. range}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

