% Auto-generated: do not edit by hand
\name{definitionTooltip}

\alias{definitionTooltip}

\title{DefinitionTooltip component}

\description{
DefinitionTooltip is a wrapper for the Carbon DefinitionTooltip component.
}

\usage{
definitionTooltip(children=NULL, id=NULL, align=NULL, autoAlign=NULL,
className=NULL, defaultOpen=NULL, definition=NULL,
loading_state=NULL, openOnHover=NULL, style=NULL,
tooltipText=NULL, triggerClassName=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{className}{Character. className}

\item{defaultOpen}{Logical | numeric | character | named list | unnamed list. defaultOpen}

\item{definition}{Logical | numeric | character | named list | unnamed list. definition}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{openOnHover}{Logical | numeric | character | named list | unnamed list. openOnHover}

\item{style}{Named list. style}

\item{tooltipText}{Logical | numeric | character | named list | unnamed list. tooltipText}

\item{triggerClassName}{Logical | numeric | character | named list | unnamed list. triggerClassName}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

