% Auto-generated: do not edit by hand
\name{dismissibleTag}

\alias{dismissibleTag}

\title{DismissibleTag component}

\description{
DismissibleTag is a wrapper for the Carbon DismissibleTag component.
}

\usage{
dismissibleTag(children=NULL, id=NULL, className=NULL, decorator=NULL,
disabled=NULL, dismissTooltipAlignment=NULL,
dismissTooltipLabel=NULL, loading_state=NULL, onClose=NULL,
renderIcon=NULL, size=NULL, slug=NULL, style=NULL,
tagTitle=NULL, text=NULL, title=NULL, type=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{dismissTooltipAlignment}{Logical | numeric | character | named list | unnamed list. dismissTooltipAlignment}

\item{dismissTooltipLabel}{Logical | numeric | character | named list | unnamed list. dismissTooltipLabel}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClose}{Logical | numeric | character | named list | unnamed list. onClose}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{tagTitle}{Logical | numeric | character | named list | unnamed list. tagTitle}

\item{text}{Logical | numeric | character | named list | unnamed list. text}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{type}{Logical | numeric | character | named list | unnamed list. type}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

