% Auto-generated: do not edit by hand
\name{dropdown}

\alias{dropdown}

\title{Dropdown component}

\description{
Dropdown is a wrapper for the Carbon Dropdown component.
}

\usage{
dropdown(children=NULL, id=NULL, ariaLabel=NULL, autoAlign=NULL,
className=NULL, decorator=NULL, direction=NULL,
disabled=NULL, downshiftProps=NULL, helperText=NULL,
hideLabel=NULL, initialSelectedItem=NULL, invalid=NULL,
invalidText=NULL, itemToElement=NULL, itemToString=NULL,
items=NULL, label=NULL, light=NULL, loading_state=NULL,
onChange=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, readOnly=NULL,
renderSelectedItem=NULL, selectedItem=NULL, size=NULL,
slug=NULL, style=NULL, title=NULL, titleText=NULL,
translateWithId=NULL, type=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{direction}{Logical | numeric | character | named list | unnamed list. direction}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{downshiftProps}{Logical | numeric | character | named list | unnamed list. downshiftProps}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{initialSelectedItem}{Logical | numeric | character | named list | unnamed list. initialSelectedItem}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{itemToElement}{Logical | numeric | character | named list | unnamed list. itemToElement}

\item{itemToString}{Logical | numeric | character | named list | unnamed list. itemToString}

\item{items}{Unnamed list. items}

\item{label}{Character. label}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{renderSelectedItem}{Logical | numeric | character | named list | unnamed list. renderSelectedItem}

\item{selectedItem}{Logical | numeric | character | named list | unnamed list. selectedItem}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{title}{Character. title}

\item{titleText}{Logical | numeric | character | named list | unnamed list. titleText}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}

\item{type}{Logical | numeric | character | named list | unnamed list. type}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

