% Auto-generated: do not edit by hand
\name{expandableTile}

\alias{expandableTile}

\title{ExpandableTile component}

\description{
ExpandableTile is a wrapper for the Carbon ExpandableTile component.
}

\usage{
expandableTile(children=NULL, id=NULL, className=NULL, decorator=NULL,
expanded=NULL, hasRoundedCorners=NULL, light=NULL,
loading_state=NULL, onClick=NULL, onKeyUp=NULL, slug=NULL,
style=NULL, tabIndex=NULL, tileCollapsedIconText=NULL,
tileCollapsedLabel=NULL, tileExpandedIconText=NULL,
tileExpandedLabel=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{expanded}{Logical. expanded}

\item{hasRoundedCorners}{Logical | numeric | character | named list | unnamed list. hasRoundedCorners}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyUp}{Logical | numeric | character | named list | unnamed list. onKeyUp}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}

\item{tileCollapsedIconText}{A list of or a singular dash component, string or number. tileCollapsedIconText}

\item{tileCollapsedLabel}{Logical | numeric | character | named list | unnamed list. tileCollapsedLabel}

\item{tileExpandedIconText}{A list of or a singular dash component, string or number. tileExpandedIconText}

\item{tileExpandedLabel}{Logical | numeric | character | named list | unnamed list. tileExpandedLabel}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

