% Auto-generated: do not edit by hand
\name{featureFlags}

\alias{featureFlags}

\title{FeatureFlags component}

\description{
FeatureFlags is a wrapper for the Carbon FeatureFlags component.
}

\usage{
featureFlags(children=NULL, id=NULL, className=NULL,
enableDialogElement=NULL, enableEnhancedFileUploader=NULL,
enableExperimentalFocusWrapWithoutSentinels=NULL,
enableFocusWrapWithoutSentinels=NULL, enablePresence=NULL,
enableTreeviewControllable=NULL,
enableV12DynamicFloatingStyles=NULL,
enableV12Overflowmenu=NULL, enableV12TileDefaultIcons=NULL,
enableV12TileRadioIcons=NULL, flags=NULL,
loading_state=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{enableDialogElement}{Logical | numeric | character | named list | unnamed list. enableDialogElement}

\item{enableEnhancedFileUploader}{Logical | numeric | character | named list | unnamed list. enableEnhancedFileUploader}

\item{enableExperimentalFocusWrapWithoutSentinels}{Logical | numeric | character | named list | unnamed list. enableExperimentalFocusWrapWithoutSentinels}

\item{enableFocusWrapWithoutSentinels}{Logical | numeric | character | named list | unnamed list. enableFocusWrapWithoutSentinels}

\item{enablePresence}{Logical | numeric | character | named list | unnamed list. enablePresence}

\item{enableTreeviewControllable}{Logical | numeric | character | named list | unnamed list. enableTreeviewControllable}

\item{enableV12DynamicFloatingStyles}{Logical | numeric | character | named list | unnamed list. enableV12DynamicFloatingStyles}

\item{enableV12Overflowmenu}{Logical | numeric | character | named list | unnamed list. enableV12Overflowmenu}

\item{enableV12TileDefaultIcons}{A list of or a singular dash component, string or number. enableV12TileDefaultIcons}

\item{enableV12TileRadioIcons}{A list of or a singular dash component, string or number. enableV12TileRadioIcons}

\item{flags}{Logical | numeric | character | named list | unnamed list. flags}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

