% Auto-generated: do not edit by hand
\name{filename}

\alias{filename}

\title{Filename component}

\description{
Filename is a wrapper for the Carbon Filename component.
}

\usage{
filename(children=NULL, id=NULL, className=NULL,
iconDescription=NULL, invalid=NULL, loading_state=NULL,
name=NULL, status=NULL, style=NULL, tabIndex=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{status}{Logical | numeric | character | named list | unnamed list. status}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

