% Auto-generated: do not edit by hand
\name{fileUploader}

\alias{fileUploader}

\title{FileUploader component}

\description{
FileUploader is a wrapper for the Carbon FileUploader component.
}

\usage{
fileUploader(children=NULL, id=NULL, accept=NULL, buttonKind=NULL,
buttonLabel=NULL, className=NULL, disabled=NULL,
filenameStatus=NULL, iconDescription=NULL,
labelDescription=NULL, labelTitle=NULL, loading_state=NULL,
maxFileSize=NULL, multiple=NULL, name=NULL, onAddFiles=NULL,
onChange=NULL, onClick=NULL, onDelete=NULL, size=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{accept}{Logical | numeric | character | named list | unnamed list. accept}

\item{buttonKind}{Logical | numeric | character | named list | unnamed list. buttonKind}

\item{buttonLabel}{Logical | numeric | character | named list | unnamed list. buttonLabel}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{filenameStatus}{Logical | numeric | character | named list | unnamed list. filenameStatus}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{labelDescription}{Logical | numeric | character | named list | unnamed list. labelDescription}

\item{labelTitle}{Logical | numeric | character | named list | unnamed list. labelTitle}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{maxFileSize}{Logical | numeric | character | named list | unnamed list. maxFileSize}

\item{multiple}{Logical | numeric | character | named list | unnamed list. multiple}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onAddFiles}{Logical | numeric | character | named list | unnamed list. onAddFiles}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onDelete}{Logical | numeric | character | named list | unnamed list. onDelete}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

