% Auto-generated: do not edit by hand
\name{fileUploaderButton}

\alias{fileUploaderButton}

\title{FileUploaderButton component}

\description{
FileUploaderButton is a wrapper for the Carbon FileUploaderButton component.
}

\usage{
fileUploaderButton(children=NULL, id=NULL, accept=NULL, buttonKind=NULL,
className=NULL, disableLabelChanges=NULL, disabled=NULL,
labelText=NULL, loading_state=NULL, multiple=NULL,
name=NULL, onChange=NULL, onClick=NULL, role=NULL,
size=NULL, style=NULL, tabIndex=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{accept}{Logical | numeric | character | named list | unnamed list. accept}

\item{buttonKind}{Logical | numeric | character | named list | unnamed list. buttonKind}

\item{className}{Character. className}

\item{disableLabelChanges}{Logical | numeric | character | named list | unnamed list. disableLabelChanges}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{multiple}{Logical | numeric | character | named list | unnamed list. multiple}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{role}{Logical | numeric | character | named list | unnamed list. role}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

