% Auto-generated: do not edit by hand
\name{fileUploaderDropContainer}

\alias{fileUploaderDropContainer}

\title{FileUploaderDropContainer component}

\description{
FileUploaderDropContainer is a wrapper for the Carbon FileUploaderDropContainer component.
}

\usage{
fileUploaderDropContainer(children=NULL, id=NULL, accept=NULL, className=NULL,
disabled=NULL, labelText=NULL, loading_state=NULL,
maxFileSize=NULL, multiple=NULL, name=NULL, onAddFiles=NULL,
onClick=NULL, pattern=NULL, role=NULL, style=NULL,
tabIndex=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{accept}{Logical | numeric | character | named list | unnamed list. accept}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{maxFileSize}{Logical | numeric | character | named list | unnamed list. maxFileSize}

\item{multiple}{Logical | numeric | character | named list | unnamed list. multiple}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onAddFiles}{Logical | numeric | character | named list | unnamed list. onAddFiles}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{pattern}{Logical | numeric | character | named list | unnamed list. pattern}

\item{role}{Logical | numeric | character | named list | unnamed list. role}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

