% Auto-generated: do not edit by hand
\name{fileUploaderItem}

\alias{fileUploaderItem}

\title{FileUploaderItem component}

\description{
FileUploaderItem is a wrapper for the Carbon FileUploaderItem component.
}

\usage{
fileUploaderItem(children=NULL, id=NULL, className=NULL, errorBody=NULL,
errorSubject=NULL, iconDescription=NULL, invalid=NULL,
loading_state=NULL, name=NULL, onDelete=NULL, size=NULL,
status=NULL, style=NULL, uuid=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{errorBody}{Logical | numeric | character | named list | unnamed list. errorBody}

\item{errorSubject}{Logical | numeric | character | named list | unnamed list. errorSubject}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onDelete}{Logical | numeric | character | named list | unnamed list. onDelete}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{status}{Logical | numeric | character | named list | unnamed list. status}

\item{style}{Named list. style}

\item{uuid}{Logical | numeric | character | named list | unnamed list. uuid}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

