% Auto-generated: do not edit by hand
\name{filterableMultiSelect}

\alias{filterableMultiSelect}

\title{FilterableMultiSelect component}

\description{
FilterableMultiSelect is a wrapper for the Carbon FilterableMultiSelect component.
}

\usage{
filterableMultiSelect(children=NULL, id=NULL, ariaLabel=NULL, autoAlign=NULL,
className=NULL, clearSelectionDescription=NULL,
clearSelectionText=NULL, compareItems=NULL, decorator=NULL,
direction=NULL, disabled=NULL, downshiftProps=NULL,
filterItems=NULL, hideLabel=NULL, initialSelectedItems=NULL,
inputProps=NULL, invalid=NULL, invalidText=NULL,
itemToElement=NULL, itemToString=NULL, items=NULL,
light=NULL, loading_state=NULL, locale=NULL, onChange=NULL,
onInputValueChange=NULL, onMenuChange=NULL, open=NULL,
placeholder=NULL, selectionFeedback=NULL, size=NULL,
slug=NULL, sortItems=NULL, style=NULL, titleText=NULL,
translateWithId=NULL, type=NULL, useTitleInItem=NULL,
warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{className}{Character. className}

\item{clearSelectionDescription}{Logical | numeric | character | named list | unnamed list. clearSelectionDescription}

\item{clearSelectionText}{Logical | numeric | character | named list | unnamed list. clearSelectionText}

\item{compareItems}{Logical | numeric | character | named list | unnamed list. compareItems}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{direction}{Logical | numeric | character | named list | unnamed list. direction}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{downshiftProps}{Logical | numeric | character | named list | unnamed list. downshiftProps}

\item{filterItems}{Logical | numeric | character | named list | unnamed list. filterItems}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{initialSelectedItems}{Logical | numeric | character | named list | unnamed list. initialSelectedItems}

\item{inputProps}{Logical | numeric | character | named list | unnamed list. inputProps}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{itemToElement}{Logical | numeric | character | named list | unnamed list. itemToElement}

\item{itemToString}{Logical | numeric | character | named list | unnamed list. itemToString}

\item{items}{Logical | numeric | character | named list | unnamed list. items}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{locale}{Logical | numeric | character | named list | unnamed list. locale}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onInputValueChange}{Logical | numeric | character | named list | unnamed list. onInputValueChange}

\item{onMenuChange}{Logical | numeric | character | named list | unnamed list. onMenuChange}

\item{open}{Logical. open}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{selectionFeedback}{Logical | numeric | character | named list | unnamed list. selectionFeedback}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{sortItems}{Logical | numeric | character | named list | unnamed list. sortItems}

\item{style}{Named list. style}

\item{titleText}{Logical | numeric | character | named list | unnamed list. titleText}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}

\item{type}{Logical | numeric | character | named list | unnamed list. type}

\item{useTitleInItem}{Logical | numeric | character | named list | unnamed list. useTitleInItem}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

