% Auto-generated: do not edit by hand
\name{fluidDatePicker}

\alias{fluidDatePicker}

\title{FluidDatePicker component}

\description{
FluidDatePicker is a wrapper for the Carbon FluidDatePicker component.
}

\usage{
fluidDatePicker(children=NULL, id=NULL, className=NULL, invalid=NULL,
invalidText=NULL, loading_state=NULL, readOnly=NULL,
style=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{style}{Named list. style}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

