% Auto-generated: do not edit by hand
\name{fluidDatePickerInput}

\alias{fluidDatePickerInput}

\title{FluidDatePickerInput component}

\description{
FluidDatePickerInput is a wrapper for the Carbon FluidDatePickerInput component.
}

\usage{
fluidDatePickerInput(children=NULL, id=NULL, className=NULL, datePickerType=NULL,
decorator=NULL, disabled=NULL, helperText=NULL,
hideLabel=NULL, invalid=NULL, invalidText=NULL,
labelText=NULL, loading_state=NULL, onChange=NULL,
onClick=NULL, pattern=NULL, placeholder=NULL, readOnly=NULL,
size=NULL, slug=NULL, style=NULL, type=NULL, warn=NULL,
warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{datePickerType}{Logical | numeric | character | named list | unnamed list. datePickerType}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{pattern}{Logical | numeric | character | named list | unnamed list. pattern}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{type}{Logical | numeric | character | named list | unnamed list. type}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

