% Auto-generated: do not edit by hand
\name{fluidDropdown}

\alias{fluidDropdown}

\title{FluidDropdown component}

\description{
FluidDropdown is a wrapper for the Carbon FluidDropdown component.
}

\usage{
fluidDropdown(children=NULL, id=NULL, className=NULL, direction=NULL,
disabled=NULL, initialSelectedItem=NULL, invalid=NULL,
invalidText=NULL, isCondensed=NULL, itemToElement=NULL,
itemToString=NULL, items=NULL, label=NULL,
loading_state=NULL, onChange=NULL, renderSelectedItem=NULL,
selectedItem=NULL, style=NULL, titleText=NULL,
translateWithId=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{direction}{Logical | numeric | character | named list | unnamed list. direction}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{initialSelectedItem}{Logical | numeric | character | named list | unnamed list. initialSelectedItem}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{isCondensed}{Logical | numeric | character | named list | unnamed list. isCondensed}

\item{itemToElement}{Logical | numeric | character | named list | unnamed list. itemToElement}

\item{itemToString}{Logical | numeric | character | named list | unnamed list. itemToString}

\item{items}{Logical | numeric | character | named list | unnamed list. items}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{renderSelectedItem}{Logical | numeric | character | named list | unnamed list. renderSelectedItem}

\item{selectedItem}{Logical | numeric | character | named list | unnamed list. selectedItem}

\item{style}{Named list. style}

\item{titleText}{Logical | numeric | character | named list | unnamed list. titleText}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

