% Auto-generated: do not edit by hand
\name{fluidMultiSelect}

\alias{fluidMultiSelect}

\title{FluidMultiSelect component}

\description{
FluidMultiSelect is a wrapper for the Carbon FluidMultiSelect component.
}

\usage{
fluidMultiSelect(children=NULL, id=NULL, className=NULL,
clearSelectionDescription=NULL, clearSelectionText=NULL,
compareItems=NULL, direction=NULL, disabled=NULL,
downshiftProps=NULL, initialSelectedItems=NULL,
invalid=NULL, invalidText=NULL, isCondensed=NULL,
isFilterable=NULL, itemToElement=NULL, itemToString=NULL,
items=NULL, label=NULL, loading_state=NULL, locale=NULL,
onChange=NULL, onInputValueChange=NULL, onMenuChange=NULL,
readOnly=NULL, selectedItems=NULL, selectionFeedback=NULL,
sortItems=NULL, style=NULL, titleText=NULL,
translateWithId=NULL, useTitleInItem=NULL, warn=NULL,
warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{clearSelectionDescription}{Logical | numeric | character | named list | unnamed list. clearSelectionDescription}

\item{clearSelectionText}{Logical | numeric | character | named list | unnamed list. clearSelectionText}

\item{compareItems}{Logical | numeric | character | named list | unnamed list. compareItems}

\item{direction}{Logical | numeric | character | named list | unnamed list. direction}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{downshiftProps}{Logical | numeric | character | named list | unnamed list. downshiftProps}

\item{initialSelectedItems}{Logical | numeric | character | named list | unnamed list. initialSelectedItems}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{isCondensed}{Logical | numeric | character | named list | unnamed list. isCondensed}

\item{isFilterable}{Logical | numeric | character | named list | unnamed list. isFilterable}

\item{itemToElement}{Logical | numeric | character | named list | unnamed list. itemToElement}

\item{itemToString}{Logical | numeric | character | named list | unnamed list. itemToString}

\item{items}{Logical | numeric | character | named list | unnamed list. items}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{locale}{Logical | numeric | character | named list | unnamed list. locale}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onInputValueChange}{Logical | numeric | character | named list | unnamed list. onInputValueChange}

\item{onMenuChange}{Logical | numeric | character | named list | unnamed list. onMenuChange}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{selectedItems}{Logical | numeric | character | named list | unnamed list. selectedItems}

\item{selectionFeedback}{Logical | numeric | character | named list | unnamed list. selectionFeedback}

\item{sortItems}{Logical | numeric | character | named list | unnamed list. sortItems}

\item{style}{Named list. style}

\item{titleText}{Logical | numeric | character | named list | unnamed list. titleText}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}

\item{useTitleInItem}{Logical | numeric | character | named list | unnamed list. useTitleInItem}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

