% Auto-generated: do not edit by hand
\name{fluidNumberInput}

\alias{fluidNumberInput}

\title{FluidNumberInput component}

\description{
FluidNumberInput is a wrapper for the Carbon FluidNumberInput component.
}

\usage{
fluidNumberInput(children=NULL, id=NULL, allowEmpty=NULL, className=NULL,
debounce=NULL, defaultValue=NULL, disableWheel=NULL,
disabled=NULL, formatOptions=NULL, iconDescription=NULL,
inputMode=NULL, invalid=NULL, invalidText=NULL, label=NULL,
loading_state=NULL, locale=NULL, max=NULL, min=NULL,
n_blur=NULL, n_submit=NULL, onChange=NULL, onClick=NULL,
onKeyUp=NULL, pattern=NULL, persisted_props=NULL,
persistence=NULL, persistence_type=NULL, readOnly=NULL,
step=NULL, style=NULL, translateWithId=NULL, type=NULL,
value=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{allowEmpty}{Logical | numeric | character | named list | unnamed list. allowEmpty}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. defaultValue}

\item{disableWheel}{Logical | numeric | character | named list | unnamed list. disableWheel}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{formatOptions}{Logical | numeric | character | named list | unnamed list. formatOptions}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{inputMode}{Logical | numeric | character | named list | unnamed list. inputMode}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{locale}{Logical | numeric | character | named list | unnamed list. locale}

\item{max}{Logical | numeric | character | named list | unnamed list. max}

\item{min}{Logical | numeric | character | named list | unnamed list. min}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyUp}{Logical | numeric | character | named list | unnamed list. onKeyUp}

\item{pattern}{Logical | numeric | character | named list | unnamed list. pattern}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{step}{Logical | numeric | character | named list | unnamed list. step}

\item{style}{Named list. style}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}

\item{type}{Logical | numeric | character | named list | unnamed list. type}

\item{value}{Logical | numeric | character | named list | unnamed list. value}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

