% Auto-generated: do not edit by hand
\name{fluidSearch}

\alias{fluidSearch}

\title{FluidSearch component}

\description{
FluidSearch is a wrapper for the Carbon FluidSearch component.
}

\usage{
fluidSearch(children=NULL, id=NULL, autoComplete=NULL, className=NULL,
closeButtonLabelText=NULL, debounce=NULL, defaultValue=NULL,
disabled=NULL, labelText=NULL, loading_state=NULL,
n_blur=NULL, n_submit=NULL, onChange=NULL, onClear=NULL,
onKeyDown=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, placeholder=NULL, role=NULL,
style=NULL, type=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{autoComplete}{Logical | numeric | character | named list | unnamed list. autoComplete}

\item{className}{Character. className}

\item{closeButtonLabelText}{Logical | numeric | character | named list | unnamed list. closeButtonLabelText}

\item{debounce}{Logical | numeric. debounce}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. defaultValue}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClear}{Logical | numeric | character | named list | unnamed list. onClear}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{role}{Logical | numeric | character | named list | unnamed list. role}

\item{style}{Named list. style}

\item{type}{Logical | numeric | character | named list | unnamed list. type}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

