% Auto-generated: do not edit by hand
\name{fluidTextInput}

\alias{fluidTextInput}

\title{FluidTextInput component}

\description{
FluidTextInput is a wrapper for the Carbon FluidTextInput component.
}

\usage{
fluidTextInput(children=NULL, id=NULL, className=NULL, debounce=NULL,
defaultValue=NULL, disabled=NULL, enableCounter=NULL,
invalid=NULL, invalidText=NULL, isPassword=NULL,
labelText=NULL, loading_state=NULL, maxCount=NULL,
n_blur=NULL, n_submit=NULL, onChange=NULL, onClick=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, placeholder=NULL, readOnly=NULL,
style=NULL, value=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. defaultValue}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{enableCounter}{Logical | numeric | character | named list | unnamed list. enableCounter}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{isPassword}{Logical | numeric | character | named list | unnamed list. isPassword}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{maxCount}{Logical | numeric | character | named list | unnamed list. maxCount}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{style}{Named list. style}

\item{value}{Logical | numeric | character | named list | unnamed list. value}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

