% Auto-generated: do not edit by hand
\name{fluidTimePickerSkeleton}

\alias{fluidTimePickerSkeleton}

\title{FluidTimePickerSkeleton component}

\description{
FluidTimePickerSkeleton is a wrapper for the Carbon FluidTimePickerSkeleton component.
}

\usage{
fluidTimePickerSkeleton(children=NULL, id=NULL, className=NULL, isOnlyTwo=NULL,
loading_state=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{isOnlyTwo}{Logical | numeric | character | named list | unnamed list. isOnlyTwo}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

