% Auto-generated: do not edit by hand
\name{formGroup}

\alias{formGroup}

\title{FormGroup component}

\description{
FormGroup is a wrapper for the Carbon FormGroup component.
}

\usage{
formGroup(children=NULL, id=NULL, className=NULL, disabled=NULL,
invalid=NULL, legendId=NULL, legendText=NULL,
loading_state=NULL, message=NULL, messageText=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{legendId}{Logical | numeric | character | named list | unnamed list. legendId}

\item{legendText}{Logical | numeric | character | named list | unnamed list. legendText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{message}{Logical | numeric | character | named list | unnamed list. message}

\item{messageText}{Logical | numeric | character | named list | unnamed list. messageText}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

