% Auto-generated: do not edit by hand
\name{grid}

\alias{grid}

\title{Grid component}

\description{
Grid is a wrapper for the Carbon Grid component.
}

\usage{
grid(children=NULL, id=NULL, align=NULL, as_=NULL,
className=NULL, condensed=NULL, fullWidth=NULL,
loading_state=NULL, narrow=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{condensed}{Logical. condensed}

\item{fullWidth}{Logical. fullWidth}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{narrow}{Logical. narrow}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

