% Auto-generated: do not edit by hand
\name{gridSettings}

\alias{gridSettings}

\title{GridSettings component}

\description{
GridSettings is a wrapper for the Carbon GridSettings component.
}

\usage{
gridSettings(children=NULL, id=NULL, className=NULL, loading_state=NULL,
mode=NULL, style=NULL, subgrid=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{mode}{Logical | numeric | character | named list | unnamed list. mode}

\item{style}{Named list. style}

\item{subgrid}{Logical | numeric | character | named list | unnamed list. subgrid}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

