% Auto-generated: do not edit by hand
\name{headerContainer}

\alias{headerContainer}

\title{HeaderContainer component}

\description{
HeaderContainer is a wrapper for the Carbon HeaderContainer component.
}

\usage{
headerContainer(children=NULL, id=NULL, className=NULL,
isSideNavExpanded=NULL, loading_state=NULL, render=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{isSideNavExpanded}{Logical | numeric | character | named list | unnamed list. isSideNavExpanded}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{render}{Logical | numeric | character | named list | unnamed list. render}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

