% Auto-generated: do not edit by hand
\name{headerGlobalAction}

\alias{headerGlobalAction}

\title{HeaderGlobalAction component}

\description{
HeaderGlobalAction is a wrapper for the Carbon HeaderGlobalAction component.
}

\usage{
headerGlobalAction(children=NULL, id=NULL, className=NULL, isActive=NULL,
loading_state=NULL, onClick=NULL, style=NULL,
tooltipAlignment=NULL, tooltipDropShadow=NULL,
tooltipHighContrast=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{isActive}{Logical | numeric | character | named list | unnamed list. isActive}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{style}{Named list. style}

\item{tooltipAlignment}{Logical | numeric | character | named list | unnamed list. tooltipAlignment}

\item{tooltipDropShadow}{Logical | numeric | character | named list | unnamed list. tooltipDropShadow}

\item{tooltipHighContrast}{Logical | numeric | character | named list | unnamed list. tooltipHighContrast}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

