% Auto-generated: do not edit by hand
\name{headerMenu}

\alias{headerMenu}

\title{HeaderMenu component}

\description{
HeaderMenu is a wrapper for the Carbon HeaderMenu component.
}

\usage{
headerMenu(children=NULL, id=NULL, className=NULL, focusRef=NULL,
isActive=NULL, isCurrentPage=NULL, loading_state=NULL,
menuLinkName=NULL, onBlur=NULL, onClick=NULL,
onKeyDown=NULL, renderMenuContent=NULL, style=NULL,
tabIndex=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{focusRef}{Logical | numeric | character | named list | unnamed list. focusRef}

\item{isActive}{Logical | numeric | character | named list | unnamed list. isActive}

\item{isCurrentPage}{Logical | numeric | character | named list | unnamed list. isCurrentPage}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{menuLinkName}{Logical | numeric | character | named list | unnamed list. menuLinkName}

\item{onBlur}{Logical | numeric | character | named list | unnamed list. onBlur}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{renderMenuContent}{Logical | numeric | character | named list | unnamed list. renderMenuContent}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

