% Auto-generated: do not edit by hand
\name{headerMenuButton}

\alias{headerMenuButton}

\title{HeaderMenuButton component}

\description{
HeaderMenuButton is a wrapper for the Carbon HeaderMenuButton component.
}

\usage{
headerMenuButton(children=NULL, id=NULL, className=NULL, isActive=NULL,
isCollapsible=NULL, loading_state=NULL, onClick=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{isActive}{Logical | numeric | character | named list | unnamed list. isActive}

\item{isCollapsible}{Logical | numeric | character | named list | unnamed list. isCollapsible}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

