% Auto-generated: do not edit by hand
\name{headerMenuItem}

\alias{headerMenuItem}

\title{HeaderMenuItem component}

\description{
HeaderMenuItem is a wrapper for the Carbon HeaderMenuItem component.
}

\usage{
headerMenuItem(children=NULL, id=NULL, as_=NULL, className=NULL,
element=NULL, isActive=NULL, isCurrentPage=NULL,
isSideNavExpanded=NULL, loading_state=NULL, role=NULL,
style=NULL, tabIndex=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{element}{Logical | numeric | character | named list | unnamed list. element}

\item{isActive}{Logical | numeric | character | named list | unnamed list. isActive}

\item{isCurrentPage}{Logical | numeric | character | named list | unnamed list. isCurrentPage}

\item{isSideNavExpanded}{Logical | numeric | character | named list | unnamed list. isSideNavExpanded}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{role}{Logical | numeric | character | named list | unnamed list. role}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

