% Auto-generated: do not edit by hand
\name{headerName}

\alias{headerName}

\title{HeaderName component}

\description{
HeaderName is a wrapper for the Carbon HeaderName component.
}

\usage{
headerName(children=NULL, id=NULL, as_=NULL, className=NULL,
element=NULL, href=NULL, isSideNavExpanded=NULL,
loading_state=NULL, prefix=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{element}{Logical | numeric | character | named list | unnamed list. element}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{isSideNavExpanded}{Logical | numeric | character | named list | unnamed list. isSideNavExpanded}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{prefix}{Logical | numeric | character | named list | unnamed list. prefix}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

