% Auto-generated: do not edit by hand
\name{headerPanel}

\alias{headerPanel}

\title{HeaderPanel component}

\description{
HeaderPanel is a wrapper for the Carbon HeaderPanel component.
}

\usage{
headerPanel(children=NULL, id=NULL, addFocusListeners=NULL,
className=NULL, expanded=NULL, href=NULL,
loading_state=NULL, onHeaderPanelFocus=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{addFocusListeners}{Logical | numeric | character | named list | unnamed list. addFocusListeners}

\item{className}{Character. className}

\item{expanded}{Logical. expanded}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onHeaderPanelFocus}{Logical | numeric | character | named list | unnamed list. onHeaderPanelFocus}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

