% Auto-generated: do not edit by hand
\name{iconButton}

\alias{iconButton}

\title{IconButton component}

\description{
IconButton is a wrapper for the Carbon IconButton component.
}

\usage{
iconButton(children=NULL, id=NULL, align=NULL, autoAlign=NULL,
badgeCount=NULL, className=NULL, closeOnActivation=NULL,
defaultOpen=NULL, disabled=NULL, dropShadow=NULL,
enterDelayMs=NULL, highContrast=NULL, href=NULL,
isSelected=NULL, kind=NULL, label=NULL, leaveDelayMs=NULL,
loading_state=NULL, rel=NULL, size=NULL, style=NULL,
target=NULL, wrapperClasses=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{badgeCount}{Logical | numeric | character | named list | unnamed list. badgeCount}

\item{className}{Character. className}

\item{closeOnActivation}{Logical | numeric | character | named list | unnamed list. closeOnActivation}

\item{defaultOpen}{Logical | numeric | character | named list | unnamed list. defaultOpen}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{dropShadow}{Logical | numeric | character | named list | unnamed list. dropShadow}

\item{enterDelayMs}{Logical | numeric | character | named list | unnamed list. enterDelayMs}

\item{highContrast}{Logical | numeric | character | named list | unnamed list. highContrast}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{isSelected}{Logical | numeric | character | named list | unnamed list. isSelected}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{leaveDelayMs}{Logical | numeric | character | named list | unnamed list. leaveDelayMs}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{rel}{Logical | numeric | character | named list | unnamed list. rel}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{target}{Logical | numeric | character | named list | unnamed list. target}

\item{wrapperClasses}{Logical | numeric | character | named list | unnamed list. wrapperClasses}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

