% Auto-generated: do not edit by hand
\name{iconIndicator}

\alias{iconIndicator}

\title{IconIndicator component}

\description{
IconIndicator is a wrapper for the Carbon IconIndicator component.
}

\usage{
iconIndicator(children=NULL, id=NULL, className=NULL, kind=NULL,
label=NULL, loading_state=NULL, size=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

