% Auto-generated: do not edit by hand
\name{iconSwitch}

\alias{iconSwitch}

\title{IconSwitch component}

\description{
IconSwitch is a wrapper for the Carbon IconSwitch component.
}

\usage{
iconSwitch(children=NULL, id=NULL, align=NULL, className=NULL,
disabled=NULL, enterDelayMs=NULL, index=NULL,
leaveDelayMs=NULL, loading_state=NULL, name=NULL,
onClick=NULL, onKeyDown=NULL, selected=NULL, size=NULL,
style=NULL, text=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{enterDelayMs}{Logical | numeric | character | named list | unnamed list. enterDelayMs}

\item{index}{Logical | numeric | character | named list | unnamed list. index}

\item{leaveDelayMs}{Logical | numeric | character | named list | unnamed list. leaveDelayMs}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{selected}{Logical | numeric | character | named list | unnamed list. selected}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{text}{Logical | numeric | character | named list | unnamed list. text}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

