% Auto-generated: do not edit by hand
\name{iconTab}

\alias{iconTab}

\title{IconTab component}

\description{
IconTab is a wrapper for the Carbon IconTab component.
}

\usage{
iconTab(children=NULL, id=NULL, badgeIndicator=NULL, className=NULL,
defaultOpen=NULL, enterDelayMs=NULL, label=NULL,
leaveDelayMs=NULL, loading_state=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{badgeIndicator}{Logical | numeric | character | named list | unnamed list. badgeIndicator}

\item{className}{Character. className}

\item{defaultOpen}{Logical | numeric | character | named list | unnamed list. defaultOpen}

\item{enterDelayMs}{Logical | numeric | character | named list | unnamed list. enterDelayMs}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{leaveDelayMs}{Logical | numeric | character | named list | unnamed list. leaveDelayMs}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

