% Auto-generated: do not edit by hand
\name{inlineLoading}

\alias{inlineLoading}

\title{InlineLoading component}

\description{
InlineLoading is a wrapper for the Carbon InlineLoading component.
}

\usage{
inlineLoading(children=NULL, id=NULL, className=NULL, description=NULL,
iconDescription=NULL, loading_state=NULL, onSuccess=NULL,
status=NULL, style=NULL, successDelay=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{description}{Logical | numeric | character | named list | unnamed list. description}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onSuccess}{Logical | numeric | character | named list | unnamed list. onSuccess}

\item{status}{Logical | numeric | character | named list | unnamed list. status}

\item{style}{Named list. style}

\item{successDelay}{Logical | numeric | character | named list | unnamed list. successDelay}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

