% Auto-generated: do not edit by hand
\name{inlineNotification}

\alias{inlineNotification}

\title{InlineNotification component}

\description{
InlineNotification is a wrapper for the Carbon InlineNotification component.
}

\usage{
inlineNotification(children=NULL, id=NULL, className=NULL,
hideCloseButton=NULL, kind=NULL, loading_state=NULL,
lowContrast=NULL, onClose=NULL, onCloseButtonClick=NULL,
role=NULL, statusIconDescription=NULL, style=NULL,
subtitle=NULL, title=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{hideCloseButton}{Logical | numeric | character | named list | unnamed list. hideCloseButton}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{lowContrast}{Logical | numeric | character | named list | unnamed list. lowContrast}

\item{onClose}{Logical | numeric | character | named list | unnamed list. onClose}

\item{onCloseButtonClick}{Logical | numeric | character | named list | unnamed list. onCloseButtonClick}

\item{role}{Logical | numeric | character | named list | unnamed list. role}

\item{statusIconDescription}{Logical | numeric | character | named list | unnamed list. statusIconDescription}

\item{style}{Named list. style}

\item{subtitle}{Logical | numeric | character | named list | unnamed list. subtitle}

\item{title}{Logical | numeric | character | named list | unnamed list. title}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

