% Auto-generated: do not edit by hand
\name{link}

\alias{link}

\title{Link component}

\description{
Link is a wrapper for the Carbon Link component.
}

\usage{
link(children=NULL, id=NULL, as_=NULL, className=NULL,
disabled=NULL, href=NULL, inline=NULL, loading_state=NULL,
renderIcon=NULL, size=NULL, style=NULL, visited=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{inline}{Logical | numeric | character | named list | unnamed list. inline}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{renderIcon}{Logical | numeric | character | named list | unnamed list. renderIcon}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{visited}{Logical | numeric | character | named list | unnamed list. visited}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

