% Auto-generated: do not edit by hand
\name{loading}

\alias{loading}

\title{Loading component}

\description{
Loading is a wrapper for the Carbon Loading component.
}

\usage{
loading(children=NULL, id=NULL, active=NULL, className=NULL,
description=NULL, loading_state=NULL, small=NULL,
style=NULL, withOverlay=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{active}{Logical. active}

\item{className}{Character. className}

\item{description}{Logical | numeric | character | named list | unnamed list. description}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{small}{Logical. small}

\item{style}{Named list. style}

\item{withOverlay}{Logical. withOverlay}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

