% Auto-generated: do not edit by hand
\name{menu}

\alias{menu}

\title{Menu component}

\description{
Menu is a wrapper for the Carbon Menu component.
}

\usage{
menu(children=NULL, id=NULL, backgroundToken=NULL, border=NULL,
className=NULL, label=NULL, loading_state=NULL,
menuAlignment=NULL, mode=NULL, onClose=NULL, onOpen=NULL,
open=NULL, size=NULL, style=NULL, target=NULL, x=NULL,
y=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{backgroundToken}{Logical | numeric | character | named list | unnamed list. backgroundToken}

\item{border}{Logical | numeric | character | named list | unnamed list. border}

\item{className}{Character. className}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{menuAlignment}{Logical | numeric | character | named list | unnamed list. menuAlignment}

\item{mode}{Logical | numeric | character | named list | unnamed list. mode}

\item{onClose}{Logical | numeric | character | named list | unnamed list. onClose}

\item{onOpen}{Logical | numeric | character | named list | unnamed list. onOpen}

\item{open}{Logical. open}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{target}{Logical | numeric | character | named list | unnamed list. target}

\item{x}{Logical | numeric | character | named list | unnamed list. x}

\item{y}{Logical | numeric | character | named list | unnamed list. y}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

