% Auto-generated: do not edit by hand
\name{menuButton}

\alias{menuButton}

\title{MenuButton component}

\description{
MenuButton is a wrapper for the Carbon MenuButton component.
}

\usage{
menuButton(children=NULL, id=NULL, className=NULL, disabled=NULL,
kind=NULL, label=NULL, loading_state=NULL,
menuAlignment=NULL, menuBackgroundToken=NULL,
menuBorder=NULL, menuTarget=NULL, size=NULL, style=NULL,
tabIndex=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{menuAlignment}{Logical | numeric | character | named list | unnamed list. menuAlignment}

\item{menuBackgroundToken}{Logical | numeric | character | named list | unnamed list. menuBackgroundToken}

\item{menuBorder}{Logical | numeric | character | named list | unnamed list. menuBorder}

\item{menuTarget}{Logical | numeric | character | named list | unnamed list. menuTarget}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

