% Auto-generated: do not edit by hand
\name{menuItem}

\alias{menuItem}

\title{MenuItem component}

\description{
MenuItem is a wrapper for the Carbon MenuItem component.
}

\usage{
menuItem(children=NULL, id=NULL, className=NULL,
dangerDescription=NULL, disabled=NULL, kind=NULL,
label=NULL, loading_state=NULL, onClick=NULL,
renderIcon=NULL, shortcut=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{dangerDescription}{Logical | numeric | character | named list | unnamed list. dangerDescription}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{renderIcon}{Logical | numeric | character | named list | unnamed list. renderIcon}

\item{shortcut}{Logical | numeric | character | named list | unnamed list. shortcut}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

