% Auto-generated: do not edit by hand
\name{menuItemRadioGroup}

\alias{menuItemRadioGroup}

\title{MenuItemRadioGroup component}

\description{
MenuItemRadioGroup is a wrapper for the Carbon MenuItemRadioGroup component.
}

\usage{
menuItemRadioGroup(children=NULL, id=NULL, className=NULL,
defaultSelectedItem=NULL, itemToString=NULL, items=NULL,
label=NULL, loading_state=NULL, onChange=NULL,
selectedItem=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultSelectedItem}{Logical | numeric | character | named list | unnamed list. defaultSelectedItem}

\item{itemToString}{Logical | numeric | character | named list | unnamed list. itemToString}

\item{items}{Logical | numeric | character | named list | unnamed list. items}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{selectedItem}{Logical | numeric | character | named list | unnamed list. selectedItem}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

