% Auto-generated: do not edit by hand
\name{menuItemSelectable}

\alias{menuItemSelectable}

\title{MenuItemSelectable component}

\description{
MenuItemSelectable is a wrapper for the Carbon MenuItemSelectable component.
}

\usage{
menuItemSelectable(children=NULL, id=NULL, className=NULL,
defaultSelected=NULL, label=NULL, loading_state=NULL,
onChange=NULL, selected=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultSelected}{Logical | numeric | character | named list | unnamed list. defaultSelected}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{selected}{Logical | numeric | character | named list | unnamed list. selected}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

