% Auto-generated: do not edit by hand
\name{modal}

\alias{modal}

\title{Modal component}

\description{
Modal is a wrapper for the Carbon Modal component.
}

\usage{
modal(children=NULL, id=NULL, alert=NULL, className=NULL,
closeButtonLabel=NULL, current=NULL, danger=NULL,
decorator=NULL, hasScrollingContent=NULL, isFullWidth=NULL,
launcherButtonRef=NULL, loadingDescription=NULL,
loadingIconDescription=NULL, loadingStatus=NULL,
loading_state=NULL, modalAriaLabel=NULL, modalHeading=NULL,
modalLabel=NULL, onKeyDown=NULL, onLoadingSuccess=NULL,
onRequestClose=NULL, onRequestSubmit=NULL,
onSecondarySubmit=NULL, open=NULL, passiveModal=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, preventCloseOnClickOutside=NULL,
primaryButtonDisabled=NULL, primaryButtonText=NULL,
secondaryButtonText=NULL, secondaryButtons=NULL,
selectorPrimaryFocus=NULL, selectorsFloatingMenus=NULL,
shouldSubmitOnEnter=NULL, size=NULL, slug=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{alert}{Logical | numeric | character | named list | unnamed list. alert}

\item{className}{Character. className}

\item{closeButtonLabel}{Logical | numeric | character | named list | unnamed list. closeButtonLabel}

\item{current}{Logical | numeric | character | named list | unnamed list. current}

\item{danger}{Logical | numeric | character | named list | unnamed list. danger}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{hasScrollingContent}{Logical | numeric | character | named list | unnamed list. hasScrollingContent}

\item{isFullWidth}{Logical | numeric | character | named list | unnamed list. isFullWidth}

\item{launcherButtonRef}{Logical | numeric | character | named list | unnamed list. launcherButtonRef}

\item{loadingDescription}{Logical | numeric | character | named list | unnamed list. loadingDescription}

\item{loadingIconDescription}{Logical | numeric | character | named list | unnamed list. loadingIconDescription}

\item{loadingStatus}{Logical | numeric | character | named list | unnamed list. loadingStatus}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{modalAriaLabel}{Logical | numeric | character | named list | unnamed list. modalAriaLabel}

\item{modalHeading}{Character. modalHeading}

\item{modalLabel}{Logical | numeric | character | named list | unnamed list. modalLabel}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{onLoadingSuccess}{Logical | numeric | character | named list | unnamed list. onLoadingSuccess}

\item{onRequestClose}{Logical | numeric | character | named list | unnamed list. onRequestClose}

\item{onRequestSubmit}{Logical | numeric | character | named list | unnamed list. onRequestSubmit}

\item{onSecondarySubmit}{Logical | numeric | character | named list | unnamed list. onSecondarySubmit}

\item{open}{Logical. open}

\item{passiveModal}{Logical | numeric | character | named list | unnamed list. passiveModal}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{preventCloseOnClickOutside}{Logical | numeric | character | named list | unnamed list. preventCloseOnClickOutside}

\item{primaryButtonDisabled}{Logical | numeric | character | named list | unnamed list. primaryButtonDisabled}

\item{primaryButtonText}{Character. primaryButtonText}

\item{secondaryButtonText}{Character. secondaryButtonText}

\item{secondaryButtons}{Logical | numeric | character | named list | unnamed list. secondaryButtons}

\item{selectorPrimaryFocus}{Logical | numeric | character | named list | unnamed list. selectorPrimaryFocus}

\item{selectorsFloatingMenus}{Logical | numeric | character | named list | unnamed list. selectorsFloatingMenus}

\item{shouldSubmitOnEnter}{Logical | numeric | character | named list | unnamed list. shouldSubmitOnEnter}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

