% Auto-generated: do not edit by hand
\name{modalBody}

\alias{modalBody}

\title{ModalBody component}

\description{
ModalBody is a wrapper for the Carbon ModalBody component.
}

\usage{
modalBody(children=NULL, id=NULL, className=NULL, hasForm=NULL,
hasScrollingContent=NULL, loading_state=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{hasForm}{Logical | numeric | character | named list | unnamed list. hasForm}

\item{hasScrollingContent}{Logical | numeric | character | named list | unnamed list. hasScrollingContent}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

