% Auto-generated: do not edit by hand
\name{modalFooter}

\alias{modalFooter}

\title{ModalFooter component}

\description{
ModalFooter is a wrapper for the Carbon ModalFooter component.
}

\usage{
modalFooter(children=NULL, id=NULL, className=NULL, closeModal=NULL,
danger=NULL, inputref=NULL, loadingDescription=NULL,
loadingIconDescription=NULL, loadingStatus=NULL,
loading_state=NULL, onLoadingSuccess=NULL,
onRequestClose=NULL, onRequestSubmit=NULL,
primaryButtonDisabled=NULL, primaryButtonText=NULL,
primaryClassName=NULL, secondaryButtonText=NULL,
secondaryButtons=NULL, secondaryClassName=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{closeModal}{Logical | numeric | character | named list | unnamed list. closeModal}

\item{danger}{Logical | numeric | character | named list | unnamed list. danger}

\item{inputref}{Logical | numeric | character | named list | unnamed list. inputref}

\item{loadingDescription}{Logical | numeric | character | named list | unnamed list. loadingDescription}

\item{loadingIconDescription}{Logical | numeric | character | named list | unnamed list. loadingIconDescription}

\item{loadingStatus}{Logical | numeric | character | named list | unnamed list. loadingStatus}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onLoadingSuccess}{Logical | numeric | character | named list | unnamed list. onLoadingSuccess}

\item{onRequestClose}{Logical | numeric | character | named list | unnamed list. onRequestClose}

\item{onRequestSubmit}{Logical | numeric | character | named list | unnamed list. onRequestSubmit}

\item{primaryButtonDisabled}{Logical | numeric | character | named list | unnamed list. primaryButtonDisabled}

\item{primaryButtonText}{Logical | numeric | character | named list | unnamed list. primaryButtonText}

\item{primaryClassName}{Logical | numeric | character | named list | unnamed list. primaryClassName}

\item{secondaryButtonText}{Logical | numeric | character | named list | unnamed list. secondaryButtonText}

\item{secondaryButtons}{Logical | numeric | character | named list | unnamed list. secondaryButtons}

\item{secondaryClassName}{Logical | numeric | character | named list | unnamed list. secondaryClassName}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

