% Auto-generated: do not edit by hand
\name{modalHeader}

\alias{modalHeader}

\title{ModalHeader component}

\description{
ModalHeader is a wrapper for the Carbon ModalHeader component.
}

\usage{
modalHeader(children=NULL, id=NULL, buttonOnClick=NULL, className=NULL,
closeClassName=NULL, closeIconClassName=NULL,
closeModal=NULL, iconDescription=NULL, label=NULL,
labelClassName=NULL, loading_state=NULL, style=NULL,
title=NULL, titleClassName=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{buttonOnClick}{Logical | numeric | character | named list | unnamed list. buttonOnClick}

\item{className}{Character. className}

\item{closeClassName}{Logical | numeric | character | named list | unnamed list. closeClassName}

\item{closeIconClassName}{Logical | numeric | character | named list | unnamed list. closeIconClassName}

\item{closeModal}{Logical | numeric | character | named list | unnamed list. closeModal}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{labelClassName}{Logical | numeric | character | named list | unnamed list. labelClassName}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{titleClassName}{Logical | numeric | character | named list | unnamed list. titleClassName}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

