% Auto-generated: do not edit by hand
\name{modalWrapper}

\alias{modalWrapper}

\title{ModalWrapper component}

\description{
ModalWrapper is a wrapper for the Carbon ModalWrapper component.
}

\usage{
modalWrapper(children=NULL, id=NULL, buttonTriggerClassName=NULL,
buttonTriggerText=NULL, className=NULL, disabled=NULL,
handleOpen=NULL, handleSubmit=NULL, loading_state=NULL,
modalBeforeContent=NULL, modalHeading=NULL, modalLabel=NULL,
modalText=NULL, onKeyDown=NULL, passiveModal=NULL,
preventCloseOnClickOutside=NULL, primaryButtonText=NULL,
renderTriggerButtonIcon=NULL, secondaryButtonText=NULL,
selectorPrimaryFocus=NULL, shouldCloseAfterSubmit=NULL,
status=NULL, style=NULL, triggerButtonIconDescription=NULL,
triggerButtonKind=NULL, withHeader=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{buttonTriggerClassName}{Logical | numeric | character | named list | unnamed list. buttonTriggerClassName}

\item{buttonTriggerText}{Logical | numeric | character | named list | unnamed list. buttonTriggerText}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{handleOpen}{Logical | numeric | character | named list | unnamed list. handleOpen}

\item{handleSubmit}{Logical | numeric | character | named list | unnamed list. handleSubmit}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{modalBeforeContent}{Logical | numeric | character | named list | unnamed list. modalBeforeContent}

\item{modalHeading}{Logical | numeric | character | named list | unnamed list. modalHeading}

\item{modalLabel}{Logical | numeric | character | named list | unnamed list. modalLabel}

\item{modalText}{Logical | numeric | character | named list | unnamed list. modalText}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{passiveModal}{Logical | numeric | character | named list | unnamed list. passiveModal}

\item{preventCloseOnClickOutside}{Logical | numeric | character | named list | unnamed list. preventCloseOnClickOutside}

\item{primaryButtonText}{Logical | numeric | character | named list | unnamed list. primaryButtonText}

\item{renderTriggerButtonIcon}{Logical | numeric | character | named list | unnamed list. renderTriggerButtonIcon}

\item{secondaryButtonText}{Logical | numeric | character | named list | unnamed list. secondaryButtonText}

\item{selectorPrimaryFocus}{Logical | numeric | character | named list | unnamed list. selectorPrimaryFocus}

\item{shouldCloseAfterSubmit}{Logical | numeric | character | named list | unnamed list. shouldCloseAfterSubmit}

\item{status}{Logical | numeric | character | named list | unnamed list. status}

\item{style}{Named list. style}

\item{triggerButtonIconDescription}{Logical | numeric | character | named list | unnamed list. triggerButtonIconDescription}

\item{triggerButtonKind}{Logical | numeric | character | named list | unnamed list. triggerButtonKind}

\item{withHeader}{Logical | numeric | character | named list | unnamed list. withHeader}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

