% Auto-generated: do not edit by hand
\name{notificationActionButton}

\alias{notificationActionButton}

\title{NotificationActionButton component}

\description{
NotificationActionButton is a wrapper for the Carbon NotificationActionButton component.
}

\usage{
notificationActionButton(children=NULL, id=NULL, className=NULL, inline=NULL,
loading_state=NULL, onClick=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{inline}{Logical | numeric | character | named list | unnamed list. inline}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

