% Auto-generated: do not edit by hand
\name{notificationButton}

\alias{notificationButton}

\title{NotificationButton component}

\description{
NotificationButton is a wrapper for the Carbon NotificationButton component.
}

\usage{
notificationButton(children=NULL, id=NULL, ariaLabel=NULL, className=NULL,
loading_state=NULL, name=NULL, notificationType=NULL,
renderIcon=NULL, style=NULL, type=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{className}{Character. className}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{notificationType}{Logical | numeric | character | named list | unnamed list. notificationType}

\item{renderIcon}{Logical | numeric | character | named list | unnamed list. renderIcon}

\item{style}{Named list. style}

\item{type}{Logical | numeric | character | named list | unnamed list. type}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

