% Auto-generated: do not edit by hand
\name{operationalTag}

\alias{operationalTag}

\title{OperationalTag component}

\description{
OperationalTag is a wrapper for the Carbon OperationalTag component.
}

\usage{
operationalTag(children=NULL, id=NULL, className=NULL, disabled=NULL,
loading_state=NULL, renderIcon=NULL, size=NULL, style=NULL,
text=NULL, type=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{text}{Logical | numeric | character | named list | unnamed list. text}

\item{type}{Logical | numeric | character | named list | unnamed list. type}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

