% Auto-generated: do not edit by hand
\name{orderedList}

\alias{orderedList}

\title{OrderedList component}

\description{
OrderedList is a wrapper for the Carbon OrderedList component.
}

\usage{
orderedList(children=NULL, id=NULL, className=NULL, isExpressive=NULL,
loading_state=NULL, native=NULL, nested=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{isExpressive}{Logical | numeric | character | named list | unnamed list. isExpressive}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{native}{Logical | numeric | character | named list | unnamed list. native}

\item{nested}{Logical | numeric | character | named list | unnamed list. nested}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

