% Auto-generated: do not edit by hand
\name{overflowMenu}

\alias{overflowMenu}

\title{OverflowMenu component}

\description{
OverflowMenu is a wrapper for the Carbon OverflowMenu component.
}

\usage{
overflowMenu(children=NULL, id=NULL, align=NULL, ariaLabel=NULL,
className=NULL, direction=NULL, flipped=NULL,
focusTrap=NULL, iconClass=NULL, iconDescription=NULL,
left=NULL, light=NULL, loading_state=NULL, menuOffset=NULL,
menuOffsetFlip=NULL, menuOptionsClass=NULL, onClick=NULL,
onClose=NULL, onFocus=NULL, onKeyDown=NULL, onOpen=NULL,
open=NULL, renderIcon=NULL, selectorPrimaryFocus=NULL,
size=NULL, style=NULL, top=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{className}{Character. className}

\item{direction}{Logical | numeric | character | named list | unnamed list. direction}

\item{flipped}{Logical | numeric | character | named list | unnamed list. flipped}

\item{focusTrap}{Logical | numeric | character | named list | unnamed list. focusTrap}

\item{iconClass}{A list of or a singular dash component, string or number. iconClass}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{left}{Logical | numeric | character | named list | unnamed list. left}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{menuOffset}{Logical | numeric | character | named list | unnamed list. menuOffset}

\item{menuOffsetFlip}{Logical | numeric | character | named list | unnamed list. menuOffsetFlip}

\item{menuOptionsClass}{Logical | numeric | character | named list | unnamed list. menuOptionsClass}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onClose}{Logical | numeric | character | named list | unnamed list. onClose}

\item{onFocus}{Logical | numeric | character | named list | unnamed list. onFocus}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{onOpen}{Logical | numeric | character | named list | unnamed list. onOpen}

\item{open}{Logical. open}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{selectorPrimaryFocus}{Logical | numeric | character | named list | unnamed list. selectorPrimaryFocus}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{top}{Logical | numeric | character | named list | unnamed list. top}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

