% Auto-generated: do not edit by hand
\name{overflowMenuItem}

\alias{overflowMenuItem}

\title{OverflowMenuItem component}

\description{
OverflowMenuItem is a wrapper for the Carbon OverflowMenuItem component.
}

\usage{
overflowMenuItem(children=NULL, id=NULL, className=NULL, closeMenu=NULL,
dangerDescription=NULL, disabled=NULL,
handleOverflowMenuItemFocus=NULL, hasDivider=NULL,
href=NULL, index=NULL, isDelete=NULL, itemText=NULL,
loading_state=NULL, onBlur=NULL, onClick=NULL, onFocus=NULL,
onKeyDown=NULL, onKeyUp=NULL, onMouseDown=NULL,
onMouseEnter=NULL, onMouseLeave=NULL, onMouseUp=NULL,
requireTitle=NULL, style=NULL, title=NULL,
wrapperClassName=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{closeMenu}{Logical | numeric | character | named list | unnamed list. closeMenu}

\item{dangerDescription}{Logical | numeric | character | named list | unnamed list. dangerDescription}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{handleOverflowMenuItemFocus}{Logical | numeric | character | named list | unnamed list. handleOverflowMenuItemFocus}

\item{hasDivider}{Logical | numeric | character | named list | unnamed list. hasDivider}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{index}{Logical | numeric | character | named list | unnamed list. index}

\item{isDelete}{Logical | numeric | character | named list | unnamed list. isDelete}

\item{itemText}{Logical | numeric | character | named list | unnamed list. itemText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onBlur}{Logical | numeric | character | named list | unnamed list. onBlur}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onFocus}{Logical | numeric | character | named list | unnamed list. onFocus}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{onKeyUp}{Logical | numeric | character | named list | unnamed list. onKeyUp}

\item{onMouseDown}{Logical | numeric | character | named list | unnamed list. onMouseDown}

\item{onMouseEnter}{Logical | numeric | character | named list | unnamed list. onMouseEnter}

\item{onMouseLeave}{Logical | numeric | character | named list | unnamed list. onMouseLeave}

\item{onMouseUp}{Logical | numeric | character | named list | unnamed list. onMouseUp}

\item{requireTitle}{Logical | numeric | character | named list | unnamed list. requireTitle}

\item{style}{Named list. style}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{wrapperClassName}{Logical | numeric | character | named list | unnamed list. wrapperClassName}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

