% Auto-generated: do not edit by hand
\name{overflowMenuV2}

\alias{overflowMenuV2}

\title{OverflowMenuV2 component}

\description{
OverflowMenuV2 is a wrapper for the Carbon OverflowMenuV2 component.
}

\usage{
overflowMenuV2(children=NULL, id=NULL, className=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The content of the component.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Specify a custom className to be applied to the component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

