% Auto-generated: do not edit by hand
\name{pagination}

\alias{pagination}

\title{Pagination component}

\description{
Pagination is a wrapper for the Carbon Pagination component.
}

\usage{
pagination(children=NULL, id=NULL, backwardText=NULL, className=NULL,
debounce=NULL, disabled=NULL, forwardText=NULL,
isLastPage=NULL, itemRangeText=NULL, itemText=NULL,
itemsPerPageText=NULL, loading_state=NULL, n_blur=NULL,
n_submit=NULL, onChange=NULL, page=NULL,
pageInputDisabled=NULL, pageNumberText=NULL,
pageRangeText=NULL, pageSelectLabelText=NULL, pageSize=NULL,
pageSizeInputDisabled=NULL, pageSizes=NULL, pageText=NULL,
pagesUnknown=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, size=NULL, style=NULL, text=NULL,
totalItems=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{backwardText}{Logical | numeric | character | named list | unnamed list. backwardText}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{forwardText}{Logical | numeric | character | named list | unnamed list. forwardText}

\item{isLastPage}{Logical | numeric | character | named list | unnamed list. isLastPage}

\item{itemRangeText}{Logical | numeric | character | named list | unnamed list. itemRangeText}

\item{itemText}{Logical | numeric | character | named list | unnamed list. itemText}

\item{itemsPerPageText}{Logical | numeric | character | named list | unnamed list. itemsPerPageText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{page}{Logical | numeric | character | named list | unnamed list. page}

\item{pageInputDisabled}{Logical | numeric | character | named list | unnamed list. pageInputDisabled}

\item{pageNumberText}{Logical | numeric | character | named list | unnamed list. pageNumberText}

\item{pageRangeText}{Logical | numeric | character | named list | unnamed list. pageRangeText}

\item{pageSelectLabelText}{Logical | numeric | character | named list | unnamed list. pageSelectLabelText}

\item{pageSize}{Logical | numeric | character | named list | unnamed list. pageSize}

\item{pageSizeInputDisabled}{Logical | numeric | character | named list | unnamed list. pageSizeInputDisabled}

\item{pageSizes}{Logical | numeric | character | named list | unnamed list. pageSizes}

\item{pageText}{Logical | numeric | character | named list | unnamed list. pageText}

\item{pagesUnknown}{Logical | numeric | character | named list | unnamed list. pagesUnknown}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{text}{Logical | numeric | character | named list | unnamed list. text}

\item{totalItems}{Logical | numeric | character | named list | unnamed list. totalItems}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

