% Auto-generated: do not edit by hand
\name{paginationNav}

\alias{paginationNav}

\title{PaginationNav component}

\description{
PaginationNav is a wrapper for the Carbon PaginationNav component.
}

\usage{
paginationNav(children=NULL, id=NULL, className=NULL,
disableOverflow=NULL, itemsShown=NULL, loading_state=NULL,
loop=NULL, onChange=NULL, page=NULL, size=NULL, style=NULL,
tooltipAlignment=NULL, tooltipPosition=NULL,
totalItems=NULL, translateWithId=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disableOverflow}{Logical | numeric | character | named list | unnamed list. disableOverflow}

\item{itemsShown}{Logical | numeric | character | named list | unnamed list. itemsShown}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{loop}{Logical | numeric | character | named list | unnamed list. loop}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{page}{Logical | numeric | character | named list | unnamed list. page}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{tooltipAlignment}{Logical | numeric | character | named list | unnamed list. tooltipAlignment}

\item{tooltipPosition}{Logical | numeric | character | named list | unnamed list. tooltipPosition}

\item{totalItems}{Logical | numeric | character | named list | unnamed list. totalItems}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

