% Auto-generated: do not edit by hand
\name{popover}

\alias{popover}

\title{Popover component}

\description{
Popover is a wrapper for the Carbon Popover component.
}

\usage{
popover(children=NULL, id=NULL, align=NULL,
alignmentAxisOffset=NULL, as_=NULL, autoAlign=NULL,
autoAlignBoundary=NULL, backgroundToken=NULL, border=NULL,
caret=NULL, className=NULL, dropShadow=NULL, height=NULL,
highContrast=NULL, isTabTip=NULL, loading_state=NULL,
onRequestClose=NULL, open=NULL, style=NULL, width=NULL,
x=NULL, y=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{alignmentAxisOffset}{Logical | numeric | character | named list | unnamed list. alignmentAxisOffset}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{autoAlignBoundary}{Logical | numeric | character | named list | unnamed list. autoAlignBoundary}

\item{backgroundToken}{Logical | numeric | character | named list | unnamed list. backgroundToken}

\item{border}{Logical | numeric | character | named list | unnamed list. border}

\item{caret}{Logical | numeric | character | named list | unnamed list. caret}

\item{className}{Character. className}

\item{dropShadow}{Logical | numeric | character | named list | unnamed list. dropShadow}

\item{height}{Logical | numeric | character | named list | unnamed list. height}

\item{highContrast}{Logical | numeric | character | named list | unnamed list. highContrast}

\item{isTabTip}{Logical | numeric | character | named list | unnamed list. isTabTip}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onRequestClose}{Logical | numeric | character | named list | unnamed list. onRequestClose}

\item{open}{Logical. open}

\item{style}{Named list. style}

\item{width}{Logical | numeric | character | named list | unnamed list. width}

\item{x}{Logical | numeric | character | named list | unnamed list. x}

\item{y}{Logical | numeric | character | named list | unnamed list. y}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

