% Auto-generated: do not edit by hand
\name{progressBar}

\alias{progressBar}

\title{ProgressBar component}

\description{
ProgressBar is a wrapper for the Carbon ProgressBar component.
}

\usage{
progressBar(children=NULL, id=NULL, className=NULL, debounce=NULL,
helperText=NULL, hideLabel=NULL, label=NULL,
loading_state=NULL, max=NULL, n_blur=NULL, n_submit=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, size=NULL, status=NULL, style=NULL,
type=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{max}{Logical | numeric | character | named list | unnamed list. max}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{status}{Logical | numeric | character | named list | unnamed list. status}

\item{style}{Named list. style}

\item{type}{Logical | numeric | character | named list | unnamed list. type}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

