% Auto-generated: do not edit by hand
\name{progressIndicator}

\alias{progressIndicator}

\title{ProgressIndicator component}

\description{
ProgressIndicator is a wrapper for the Carbon ProgressIndicator component.
}

\usage{
progressIndicator(children=NULL, id=NULL, className=NULL, currentIndex=NULL,
loading_state=NULL, onChange=NULL, spaceEqually=NULL,
style=NULL, vertical=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{currentIndex}{Numeric. currentIndex}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{spaceEqually}{Logical. spaceEqually}

\item{style}{Named list. style}

\item{vertical}{Logical. vertical}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

