% Auto-generated: do not edit by hand
\name{progressIndicatorSkeleton}

\alias{progressIndicatorSkeleton}

\title{ProgressIndicatorSkeleton component}

\description{
ProgressIndicatorSkeleton is a wrapper for the Carbon ProgressIndicatorSkeleton component.
}

\usage{
progressIndicatorSkeleton(children=NULL, id=NULL, className=NULL, loading_state=NULL,
style=NULL, vertical=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}

\item{vertical}{Logical | numeric | character | named list | unnamed list. vertical}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

