% Auto-generated: do not edit by hand
\name{progressStep}

\alias{progressStep}

\title{ProgressStep component}

\description{
ProgressStep is a wrapper for the Carbon ProgressStep component.
}

\usage{
progressStep(children=NULL, id=NULL, className=NULL, complete=NULL,
current=NULL, description=NULL, disabled=NULL, index=NULL,
invalid=NULL, label=NULL, loading_state=NULL, onClick=NULL,
overflowTooltipProps=NULL, secondaryLabel=NULL, style=NULL,
tooltipId=NULL, translateWithId=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{complete}{Logical | numeric | character | named list | unnamed list. complete}

\item{current}{Logical | numeric | character | named list | unnamed list. current}

\item{description}{Character. description}

\item{disabled}{Logical. disabled}

\item{index}{Logical | numeric | character | named list | unnamed list. index}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{label}{A list of or a singular dash component, string or number. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{overflowTooltipProps}{Logical | numeric | character | named list | unnamed list. overflowTooltipProps}

\item{secondaryLabel}{Character. secondaryLabel}

\item{style}{Named list. style}

\item{tooltipId}{Logical | numeric | character | named list | unnamed list. tooltipId}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

