% Auto-generated: do not edit by hand
\name{radioButton}

\alias{radioButton}

\title{RadioButton component}

\description{
RadioButton is a wrapper for the Carbon RadioButton component.
}

\usage{
radioButton(children=NULL, id=NULL, checked=NULL, className=NULL,
debounce=NULL, decorator=NULL, defaultChecked=NULL,
disabled=NULL, hideLabel=NULL, invalid=NULL,
invalidText=NULL, label=NULL, labelPosition=NULL,
labelText=NULL, loading_state=NULL, n_blur=NULL,
n_submit=NULL, name=NULL, onChange=NULL, onClick=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, readOnly=NULL, required=NULL,
slug=NULL, style=NULL, value=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{checked}{Logical. checked}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{defaultChecked}{Logical | numeric | character | named list | unnamed list. defaultChecked}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{label}{Character. label}

\item{labelPosition}{Logical | numeric | character | named list | unnamed list. labelPosition}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{required}{Logical | numeric | character | named list | unnamed list. required}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{value}{Logical | numeric | character | named list | unnamed list. value}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

