% Auto-generated: do not edit by hand
\name{radioButtonGroup}

\alias{radioButtonGroup}

\title{RadioButtonGroup component}

\description{
RadioButtonGroup is a wrapper for the Carbon RadioButtonGroup component.
}

\usage{
radioButtonGroup(children=NULL, id=NULL, className=NULL, decorator=NULL,
defaultSelected=NULL, disabled=NULL, helperText=NULL,
invalid=NULL, invalidText=NULL, label=NULL,
labelPosition=NULL, legendText=NULL, loading_state=NULL,
name=NULL, onChange=NULL, orientation=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, readOnly=NULL, required=NULL,
slug=NULL, style=NULL, valueSelected=NULL, warn=NULL,
warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{defaultSelected}{Logical | numeric | character | named list | unnamed list. defaultSelected}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{label}{Character. label}

\item{labelPosition}{Logical | numeric | character | named list | unnamed list. labelPosition}

\item{legendText}{Logical | numeric | character | named list | unnamed list. legendText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{orientation}{Logical | numeric | character | named list | unnamed list. orientation}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{required}{Logical | numeric | character | named list | unnamed list. required}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{valueSelected}{Logical | numeric | character | named list | unnamed list. valueSelected}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

