% Auto-generated: do not edit by hand
\name{radioTile}

\alias{radioTile}

\title{RadioTile component}

\description{
RadioTile is a wrapper for the Carbon RadioTile component.
}

\usage{
radioTile(children=NULL, id=NULL, checked=NULL, className=NULL,
debounce=NULL, decorator=NULL, disabled=NULL, event=NULL,
hasRoundedCorners=NULL, light=NULL, loading_state=NULL,
n_blur=NULL, n_submit=NULL, name=NULL, onChange=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, required=NULL, slug=NULL, style=NULL,
tabIndex=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{checked}{Logical | numeric | character | named list | unnamed list. checked}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{event}{Logical | numeric | character | named list | unnamed list. event}

\item{hasRoundedCorners}{Logical | numeric | character | named list | unnamed list. hasRoundedCorners}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{required}{Logical | numeric | character | named list | unnamed list. required}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

