% Auto-generated: do not edit by hand
\name{row}

\alias{row}

\title{Row component}

\description{
Row is a wrapper for the Carbon Row component.
}

\usage{
row(children=NULL, id=NULL, as_=NULL, className=NULL,
condensed=NULL, loading_state=NULL, narrow=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{condensed}{Logical. condensed}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{narrow}{Logical. narrow}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

