% Auto-generated: do not edit by hand
\name{select}

\alias{select}

\title{Select component}

\description{
Select is a wrapper for the Carbon Select component.
}

\usage{
select(children=NULL, id=NULL, className=NULL, decorator=NULL,
defaultValue=NULL, disabled=NULL, helperText=NULL,
hideLabel=NULL, inline=NULL, invalid=NULL, invalidText=NULL,
label=NULL, labelText=NULL, light=NULL, loading_state=NULL,
noLabel=NULL, onChange=NULL, persisted_props=NULL,
persistence=NULL, persistence_type=NULL, readOnly=NULL,
size=NULL, slug=NULL, style=NULL, value=NULL, warn=NULL,
warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. defaultValue}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{inline}{Logical | numeric | character | named list | unnamed list. inline}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{label}{Character. label}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{noLabel}{Logical | numeric | character | named list | unnamed list. noLabel}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{value}{Logical | numeric | character | named list | unnamed list. value}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

