% Auto-generated: do not edit by hand
\name{selectableTag}

\alias{selectableTag}

\title{SelectableTag component}

\description{
SelectableTag is a wrapper for the Carbon SelectableTag component.
}

\usage{
selectableTag(children=NULL, id=NULL, className=NULL,
defaultSelected=NULL, disabled=NULL, loading_state=NULL,
onChange=NULL, onClick=NULL, renderIcon=NULL, selected=NULL,
size=NULL, style=NULL, text=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultSelected}{Logical | numeric | character | named list | unnamed list. defaultSelected}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{renderIcon}{Logical | numeric | character | named list | unnamed list. renderIcon}

\item{selected}{Logical | numeric | character | named list | unnamed list. selected}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{text}{Logical | numeric | character | named list | unnamed list. text}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

