% Auto-generated: do not edit by hand
\name{selectableTile}

\alias{selectableTile}

\title{SelectableTile component}

\description{
SelectableTile is a wrapper for the Carbon SelectableTile component.
}

\usage{
selectableTile(children=NULL, id=NULL, className=NULL, debounce=NULL,
decorator=NULL, disabled=NULL, hasRoundedCorners=NULL,
light=NULL, loading_state=NULL, n_blur=NULL, n_submit=NULL,
name=NULL, onChange=NULL, onClick=NULL, onKeyDown=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, selected=NULL, slug=NULL, style=NULL,
tabIndex=NULL, title=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{hasRoundedCorners}{Logical | numeric | character | named list | unnamed list. hasRoundedCorners}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{selected}{Logical | numeric | character | named list | unnamed list. selected}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

