% Auto-generated: do not edit by hand
\name{selectItem}

\alias{selectItem}

\title{SelectItem component}

\description{
SelectItem is a wrapper for the Carbon SelectItem component.
}

\usage{
selectItem(children=NULL, id=NULL, className=NULL, debounce=NULL,
disabled=NULL, hidden=NULL, loading_state=NULL, n_blur=NULL,
n_submit=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, style=NULL, text=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{hidden}{Logical | numeric | character | named list | unnamed list. hidden}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{style}{Named list. style}

\item{text}{Character. text}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

