% Auto-generated: do not edit by hand
\name{selectItemGroup}

\alias{selectItemGroup}

\title{SelectItemGroup component}

\description{
SelectItemGroup is a wrapper for the Carbon SelectItemGroup component.
}

\usage{
selectItemGroup(children=NULL, id=NULL, className=NULL, disabled=NULL,
label=NULL, loading_state=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

