% Auto-generated: do not edit by hand
\name{sideNav}

\alias{sideNav}

\title{SideNav component}

\description{
SideNav is a wrapper for the Carbon SideNav component.
}

\usage{
sideNav(children=NULL, id=NULL, addFocusListeners=NULL,
addMouseListeners=NULL, className=NULL,
defaultExpanded=NULL, enterDelayMs=NULL, expanded=NULL,
href=NULL, isChildOfHeader=NULL, isFixedNav=NULL,
isPersistent=NULL, isRail=NULL, loading_state=NULL,
onOverlayClick=NULL, onSideNavBlur=NULL, onToggle=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{addFocusListeners}{Logical | numeric | character | named list | unnamed list. addFocusListeners}

\item{addMouseListeners}{Logical | numeric | character | named list | unnamed list. addMouseListeners}

\item{className}{Character. className}

\item{defaultExpanded}{Logical | numeric | character | named list | unnamed list. defaultExpanded}

\item{enterDelayMs}{Logical | numeric | character | named list | unnamed list. enterDelayMs}

\item{expanded}{Logical | numeric | character | named list | unnamed list. expanded}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{isChildOfHeader}{Logical | numeric | character | named list | unnamed list. isChildOfHeader}

\item{isFixedNav}{Logical | numeric | character | named list | unnamed list. isFixedNav}

\item{isPersistent}{Logical | numeric | character | named list | unnamed list. isPersistent}

\item{isRail}{Logical | numeric | character | named list | unnamed list. isRail}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onOverlayClick}{Logical | numeric | character | named list | unnamed list. onOverlayClick}

\item{onSideNavBlur}{Logical | numeric | character | named list | unnamed list. onSideNavBlur}

\item{onToggle}{Logical | numeric | character | named list | unnamed list. onToggle}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

