% Auto-generated: do not edit by hand
\name{sideNavDetails}

\alias{sideNavDetails}

\title{SideNavDetails component}

\description{
SideNavDetails is a wrapper for the Carbon SideNavDetails component.
}

\usage{
sideNavDetails(children=NULL, id=NULL, className=NULL, loading_state=NULL,
style=NULL, title=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}

\item{title}{Logical | numeric | character | named list | unnamed list. title}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

