% Auto-generated: do not edit by hand
\name{sideNavFooter}

\alias{sideNavFooter}

\title{SideNavFooter component}

\description{
SideNavFooter is a wrapper for the Carbon SideNavFooter component.
}

\usage{
sideNavFooter(children=NULL, id=NULL, assistiveText=NULL, className=NULL,
expanded=NULL, loading_state=NULL, onToggle=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{assistiveText}{Logical | numeric | character | named list | unnamed list. assistiveText}

\item{className}{Character. className}

\item{expanded}{Logical | numeric | character | named list | unnamed list. expanded}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onToggle}{Logical | numeric | character | named list | unnamed list. onToggle}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

