% Auto-generated: do not edit by hand
\name{sideNavMenu}

\alias{sideNavMenu}

\title{SideNavMenu component}

\description{
SideNavMenu is a wrapper for the Carbon SideNavMenu component.
}

\usage{
sideNavMenu(children=NULL, id=NULL, className=NULL,
defaultExpanded=NULL, isActive=NULL, isSideNavExpanded=NULL,
large=NULL, loading_state=NULL, renderIcon=NULL, style=NULL,
tabIndex=NULL, title=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultExpanded}{Logical | numeric | character | named list | unnamed list. defaultExpanded}

\item{isActive}{Logical | numeric | character | named list | unnamed list. isActive}

\item{isSideNavExpanded}{Logical | numeric | character | named list | unnamed list. isSideNavExpanded}

\item{large}{Logical | numeric | character | named list | unnamed list. large}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{renderIcon}{Logical | numeric | character | named list | unnamed list. renderIcon}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}

\item{title}{Logical | numeric | character | named list | unnamed list. title}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

