% Auto-generated: do not edit by hand
\name{sideNavSwitcher}

\alias{sideNavSwitcher}

\title{SideNavSwitcher component}

\description{
SideNavSwitcher is a wrapper for the Carbon SideNavSwitcher component.
}

\usage{
sideNavSwitcher(children=NULL, id=NULL, className=NULL, labelText=NULL,
loading_state=NULL, onChange=NULL, options=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{options}{Logical | numeric | character | named list | unnamed list. options}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

