% Auto-generated: do not edit by hand
\name{skipToContent}

\alias{skipToContent}

\title{SkipToContent component}

\description{
SkipToContent is a wrapper for the Carbon SkipToContent component.
}

\usage{
skipToContent(children=NULL, id=NULL, className=NULL, href=NULL,
loading_state=NULL, style=NULL, tabIndex=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

