% Auto-generated: do not edit by hand
\name{slider}

\alias{slider}

\title{Slider component}

\description{
Slider is a wrapper for the Carbon Slider component.
}

\usage{
slider(children=NULL, id=NULL, ariaLabelInput=NULL, className=NULL,
debounce=NULL, disabled=NULL, formatLabel=NULL,
hideLabel=NULL, hideTextInput=NULL, inputType=NULL,
invalid=NULL, invalidText=NULL, label=NULL, labelText=NULL,
light=NULL, loading_state=NULL, max=NULL, maxLabel=NULL,
min=NULL, minLabel=NULL, n_blur=NULL, n_submit=NULL,
name=NULL, onBlur=NULL, onChange=NULL, onInputKeyUp=NULL,
onRelease=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, readOnly=NULL, required=NULL,
step=NULL, stepMultiplier=NULL, style=NULL,
translateWithId=NULL, unstable_ariaLabelInputUpper=NULL,
unstable_nameUpper=NULL, unstable_valueUpper=NULL,
value=NULL, warn=NULL, warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabelInput}{Logical | numeric | character | named list | unnamed list. ariaLabelInput}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{formatLabel}{Logical | numeric | character | named list | unnamed list. formatLabel}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{hideTextInput}{Logical | numeric | character | named list | unnamed list. hideTextInput}

\item{inputType}{Logical | numeric | character | named list | unnamed list. inputType}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{label}{Character. label}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{max}{Numeric. max}

\item{maxLabel}{Logical | numeric | character | named list | unnamed list. maxLabel}

\item{min}{Numeric. min}

\item{minLabel}{Logical | numeric | character | named list | unnamed list. minLabel}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onBlur}{Logical | numeric | character | named list | unnamed list. onBlur}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onInputKeyUp}{Logical | numeric | character | named list | unnamed list. onInputKeyUp}

\item{onRelease}{Logical | numeric | character | named list | unnamed list. onRelease}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{required}{Logical | numeric | character | named list | unnamed list. required}

\item{step}{Logical | numeric | character | named list | unnamed list. step}

\item{stepMultiplier}{Logical | numeric | character | named list | unnamed list. stepMultiplier}

\item{style}{Named list. style}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}

\item{unstable_ariaLabelInputUpper}{Logical | numeric | character | named list | unnamed list. unstable_ariaLabelInputUpper}

\item{unstable_nameUpper}{Logical | numeric | character | named list | unnamed list. unstable_nameUpper}

\item{unstable_valueUpper}{Logical | numeric | character | named list | unnamed list. unstable_valueUpper}

\item{value}{Numeric. value}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

