% Auto-generated: do not edit by hand
\name{sliderSkeleton}

\alias{sliderSkeleton}

\title{SliderSkeleton component}

\description{
SliderSkeleton is a wrapper for the Carbon SliderSkeleton component.
}

\usage{
sliderSkeleton(children=NULL, id=NULL, ariaLabel=NULL, className=NULL,
hideLabel=NULL, loading_state=NULL, style=NULL,
twoHandles=NULL, unstable_ariaLabelHandleUpper=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{className}{Character. className}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}

\item{twoHandles}{Logical | numeric | character | named list | unnamed list. twoHandles}

\item{unstable_ariaLabelHandleUpper}{Logical | numeric | character | named list | unnamed list. unstable_ariaLabelHandleUpper}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

