% Auto-generated: do not edit by hand
\name{slugContent}

\alias{slugContent}

\title{SlugContent component}

\description{
SlugContent is a wrapper for the Carbon SlugContent component.
}

\usage{
slugContent(children=NULL, id=NULL, className=NULL, loading_state=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

