% Auto-generated: do not edit by hand
\name{structuredListBody}

\alias{structuredListBody}

\title{StructuredListBody component}

\description{
StructuredListBody is a wrapper for the Carbon StructuredListBody component.
}

\usage{
structuredListBody(children=NULL, id=NULL, className=NULL, head=NULL,
loading_state=NULL, onKeyDown=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{head}{Logical | numeric | character | named list | unnamed list. head}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

