% Auto-generated: do not edit by hand
\name{structuredListCell}

\alias{structuredListCell}

\title{StructuredListCell component}

\description{
StructuredListCell is a wrapper for the Carbon StructuredListCell component.
}

\usage{
structuredListCell(children=NULL, id=NULL, className=NULL, head=NULL,
loading_state=NULL, noWrap=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{head}{Logical | numeric | character | named list | unnamed list. head}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{noWrap}{Logical | numeric | character | named list | unnamed list. noWrap}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

