% Auto-generated: do not edit by hand
\name{structuredListInput}

\alias{structuredListInput}

\title{StructuredListInput component}

\description{
StructuredListInput is a wrapper for the Carbon StructuredListInput component.
}

\usage{
structuredListInput(children=NULL, id=NULL, className=NULL, debounce=NULL,
defaultChecked=NULL, loading_state=NULL, n_blur=NULL,
n_submit=NULL, name=NULL, onChange=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, style=NULL, title=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{defaultChecked}{Logical | numeric | character | named list | unnamed list. defaultChecked}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{style}{Named list. style}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

