% Auto-generated: do not edit by hand
\name{structuredListRow}

\alias{structuredListRow}

\title{StructuredListRow component}

\description{
StructuredListRow is a wrapper for the Carbon StructuredListRow component.
}

\usage{
structuredListRow(children=NULL, id=NULL, className=NULL, head=NULL,
label=NULL, loading_state=NULL, onClick=NULL,
onKeyDown=NULL, selection=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{head}{Logical | numeric | character | named list | unnamed list. head}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{selection}{Logical | numeric | character | named list | unnamed list. selection}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

