% Auto-generated: do not edit by hand
\name{structuredListWrapper}

\alias{structuredListWrapper}

\title{StructuredListWrapper component}

\description{
StructuredListWrapper is a wrapper for the Carbon StructuredListWrapper component.
}

\usage{
structuredListWrapper(children=NULL, id=NULL, ariaLabel=NULL, className=NULL,
isCondensed=NULL, isFlush=NULL, loading_state=NULL,
selectedInitialRow=NULL, selection=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{className}{Character. className}

\item{isCondensed}{Logical | numeric | character | named list | unnamed list. isCondensed}

\item{isFlush}{Logical | numeric | character | named list | unnamed list. isFlush}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{selectedInitialRow}{Logical | numeric | character | named list | unnamed list. selectedInitialRow}

\item{selection}{Logical | numeric | character | named list | unnamed list. selection}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

