% Auto-generated: do not edit by hand
\name{switch}

\alias{switch}

\title{Switch component}

\description{
Switch is a wrapper for the Carbon Switch component.
}

\usage{
switch(children=NULL, id=NULL, className=NULL, disabled=NULL,
index=NULL, loading_state=NULL, name=NULL, onClick=NULL,
onKeyDown=NULL, selected=NULL, style=NULL, text=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{index}{Logical | numeric | character | named list | unnamed list. index}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{selected}{Logical | numeric | character | named list | unnamed list. selected}

\item{style}{Named list. style}

\item{text}{Logical | numeric | character | named list | unnamed list. text}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

