% Auto-generated: do not edit by hand
\name{switcherItem}

\alias{switcherItem}

\title{SwitcherItem component}

\description{
SwitcherItem is a wrapper for the Carbon SwitcherItem component.
}

\usage{
switcherItem(children=NULL, id=NULL, className=NULL,
handleSwitcherItemFocus=NULL, href=NULL, index=NULL,
loading_state=NULL, onClick=NULL, onKeyDown=NULL, rel=NULL,
style=NULL, tabIndex=NULL, target=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{handleSwitcherItemFocus}{Logical | numeric | character | named list | unnamed list. handleSwitcherItemFocus}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{index}{Logical | numeric | character | named list | unnamed list. index}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{rel}{Logical | numeric | character | named list | unnamed list. rel}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}

\item{target}{Logical | numeric | character | named list | unnamed list. target}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

