% Auto-generated: do not edit by hand
\name{tab}

\alias{tab}

\title{Tab component}

\description{
Tab is a wrapper for the Carbon Tab component.
}

\usage{
tab(children=NULL, id=NULL, as_=NULL, className=NULL,
disabled=NULL, label=NULL, loading_state=NULL, onClick=NULL,
onKeyDown=NULL, renderButton=NULL, renderIcon=NULL,
secondaryLabel=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{disabled}{Logical. disabled}

\item{label}{A list of or a singular dash component, string or number. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{renderButton}{Logical | numeric | character | named list | unnamed list. renderButton}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{secondaryLabel}{Logical | numeric | character | named list | unnamed list. secondaryLabel}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

