% Auto-generated: do not edit by hand
\name{table}

\alias{table}

\title{Table component}

\description{
Table is a wrapper for the Carbon Table component.
}

\usage{
table(children=NULL, id=NULL, className=NULL,
experimentalAutoAlign=NULL, isSortable=NULL,
loading_state=NULL, overflowMenuOnHover=NULL, size=NULL,
stickyHeader=NULL, style=NULL, tabIndex=NULL,
useStaticWidth=NULL, useZebraStyles=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{experimentalAutoAlign}{Logical | numeric | character | named list | unnamed list. experimentalAutoAlign}

\item{isSortable}{Logical | numeric | character | named list | unnamed list. isSortable}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{overflowMenuOnHover}{Logical | numeric | character | named list | unnamed list. overflowMenuOnHover}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{stickyHeader}{Logical | numeric | character | named list | unnamed list. stickyHeader}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}

\item{useStaticWidth}{Logical | numeric | character | named list | unnamed list. useStaticWidth}

\item{useZebraStyles}{Logical | numeric | character | named list | unnamed list. useZebraStyles}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

