% Auto-generated: do not edit by hand
\name{tableBatchAction}

\alias{tableBatchAction}

\title{TableBatchAction component}

\description{
TableBatchAction is a wrapper for the Carbon TableBatchAction component.
}

\usage{
tableBatchAction(children=NULL, id=NULL, className=NULL, hasIconOnly=NULL,
iconDescription=NULL, loading_state=NULL, renderIcon=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{hasIconOnly}{Logical | numeric | character | named list | unnamed list. hasIconOnly}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

