% Auto-generated: do not edit by hand
\name{tableBatchActions}

\alias{tableBatchActions}

\title{TableBatchActions component}

\description{
TableBatchActions is a wrapper for the Carbon TableBatchActions component.
}

\usage{
tableBatchActions(children=NULL, id=NULL, className=NULL, loading_state=NULL,
onCancel=NULL, onSelectAll=NULL,
shouldShowBatchActions=NULL, style=NULL, totalCount=NULL,
totalSelected=NULL, translateWithId=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onCancel}{Logical | numeric | character | named list | unnamed list. onCancel}

\item{onSelectAll}{Logical | numeric | character | named list | unnamed list. onSelectAll}

\item{shouldShowBatchActions}{Logical | numeric | character | named list | unnamed list. shouldShowBatchActions}

\item{style}{Named list. style}

\item{totalCount}{Logical | numeric | character | named list | unnamed list. totalCount}

\item{totalSelected}{Logical | numeric | character | named list | unnamed list. totalSelected}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

