% Auto-generated: do not edit by hand
\name{tableCell}

\alias{tableCell}

\title{TableCell component}

\description{
TableCell is a wrapper for the Carbon TableCell component.
}

\usage{
tableCell(children=NULL, id=NULL, className=NULL, colSpan=NULL,
hasAILabelHeader=NULL, headers=NULL, loading_state=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{colSpan}{Logical | numeric | character | named list | unnamed list. colSpan}

\item{hasAILabelHeader}{Logical | numeric | character | named list | unnamed list. hasAILabelHeader}

\item{headers}{Logical | numeric | character | named list | unnamed list. headers}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

