% Auto-generated: do not edit by hand
\name{tableContainer}

\alias{tableContainer}

\title{TableContainer component}

\description{
TableContainer is a wrapper for the Carbon TableContainer component.
}

\usage{
tableContainer(children=NULL, id=NULL, aiEnabled=NULL, className=NULL,
decorator=NULL, description=NULL, loading_state=NULL,
stickyHeader=NULL, style=NULL, title=NULL,
useStaticWidth=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{aiEnabled}{Logical | numeric | character | named list | unnamed list. aiEnabled}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{description}{Logical | numeric | character | named list | unnamed list. description}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{stickyHeader}{Logical | numeric | character | named list | unnamed list. stickyHeader}

\item{style}{Named list. style}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{useStaticWidth}{Logical | numeric | character | named list | unnamed list. useStaticWidth}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

