% Auto-generated: do not edit by hand
\name{tableExpandRow}

\alias{tableExpandRow}

\title{TableExpandRow component}

\description{
TableExpandRow is a wrapper for the Carbon TableExpandRow component.
}

\usage{
tableExpandRow(children=NULL, id=NULL, ariaLabel=NULL, className=NULL,
expandHeader=NULL, expandIconDescription=NULL,
isExpanded=NULL, isSelected=NULL, loading_state=NULL,
onExpand=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{className}{Character. className}

\item{expandHeader}{Logical | numeric | character | named list | unnamed list. expandHeader}

\item{expandIconDescription}{Logical | numeric | character | named list | unnamed list. expandIconDescription}

\item{isExpanded}{Logical. isExpanded}

\item{isSelected}{Logical | numeric | character | named list | unnamed list. isSelected}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onExpand}{Logical | numeric | character | named list | unnamed list. onExpand}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

