% Auto-generated: do not edit by hand
\name{tableHeader}

\alias{tableHeader}

\title{TableHeader component}

\description{
TableHeader is a wrapper for the Carbon TableHeader component.
}

\usage{
tableHeader(children=NULL, id=NULL, className=NULL, colSpan=NULL,
isSortHeader=NULL, isSortable=NULL, loading_state=NULL,
onClick=NULL, scope=NULL, sortDirection=NULL, style=NULL,
translateWithId=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{colSpan}{Logical | numeric | character | named list | unnamed list. colSpan}

\item{isSortHeader}{Logical | numeric | character | named list | unnamed list. isSortHeader}

\item{isSortable}{Logical | numeric | character | named list | unnamed list. isSortable}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{scope}{Logical | numeric | character | named list | unnamed list. scope}

\item{sortDirection}{Logical | numeric | character | named list | unnamed list. sortDirection}

\item{style}{Named list. style}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

