% Auto-generated: do not edit by hand
\name{tableRow}

\alias{tableRow}

\title{TableRow component}

\description{
TableRow is a wrapper for the Carbon TableRow component.
}

\usage{
tableRow(children=NULL, id=NULL, ariaLabel=NULL, className=NULL,
isExpanded=NULL, isSelected=NULL, loading_state=NULL,
onExpand=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{className}{Character. className}

\item{isExpanded}{Logical | numeric | character | named list | unnamed list. isExpanded}

\item{isSelected}{Logical | numeric | character | named list | unnamed list. isSelected}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onExpand}{Logical | numeric | character | named list | unnamed list. onExpand}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

