% Auto-generated: do not edit by hand
\name{tableSelectAll}

\alias{tableSelectAll}

\title{TableSelectAll component}

\description{
TableSelectAll is a wrapper for the Carbon TableSelectAll component.
}

\usage{
tableSelectAll(children=NULL, id=NULL, ariaLabel=NULL, checked=NULL,
className=NULL, debounce=NULL, disabled=NULL,
indeterminate=NULL, loading_state=NULL, n_blur=NULL,
n_submit=NULL, name=NULL, onSelect=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{checked}{Logical | numeric | character | named list | unnamed list. checked}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{indeterminate}{Logical | numeric | character | named list | unnamed list. indeterminate}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onSelect}{Logical | numeric | character | named list | unnamed list. onSelect}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

