% Auto-generated: do not edit by hand
\name{tableSelectRow}

\alias{tableSelectRow}

\title{TableSelectRow component}

\description{
TableSelectRow is a wrapper for the Carbon TableSelectRow component.
}

\usage{
tableSelectRow(children=NULL, id=NULL, ariaLabel=NULL, checked=NULL,
className=NULL, debounce=NULL, disabled=NULL,
loading_state=NULL, n_blur=NULL, n_submit=NULL, name=NULL,
onChange=NULL, onSelect=NULL, persisted_props=NULL,
persistence=NULL, persistence_type=NULL, radio=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{checked}{Logical. checked}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onSelect}{Logical | numeric | character | named list | unnamed list. onSelect}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{radio}{Logical | numeric | character | named list | unnamed list. radio}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

