% Auto-generated: do not edit by hand
\name{tableToolbarMenu}

\alias{tableToolbarMenu}

\title{TableToolbarMenu component}

\description{
TableToolbarMenu is a wrapper for the Carbon TableToolbarMenu component.
}

\usage{
tableToolbarMenu(children=NULL, id=NULL, className=NULL,
iconDescription=NULL, loading_state=NULL,
menuOptionsClass=NULL, renderIcon=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{iconDescription}{Logical | numeric | character | named list | unnamed list. iconDescription}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{menuOptionsClass}{Logical | numeric | character | named list | unnamed list. menuOptionsClass}

\item{renderIcon}{Logical | numeric | character | named list | unnamed list. renderIcon}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

