% Auto-generated: do not edit by hand
\name{tableToolbarSearch}

\alias{tableToolbarSearch}

\title{TableToolbarSearch component}

\description{
TableToolbarSearch is a wrapper for the Carbon TableToolbarSearch component.
}

\usage{
tableToolbarSearch(children=NULL, id=NULL, className=NULL,
defaultExpanded=NULL, defaultValue=NULL, disabled=NULL,
expanded=NULL, labelText=NULL, loading_state=NULL,
onBlur=NULL, onChange=NULL, onClear=NULL, onExpand=NULL,
onFocus=NULL, persistent=NULL, placeholder=NULL,
searchContainerClass=NULL, size=NULL, style=NULL,
tabIndex=NULL, translateWithId=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultExpanded}{Logical | numeric | character | named list | unnamed list. defaultExpanded}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. defaultValue}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{expanded}{Logical. expanded}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onBlur}{Logical | numeric | character | named list | unnamed list. onBlur}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClear}{Logical | numeric | character | named list | unnamed list. onClear}

\item{onExpand}{Logical | numeric | character | named list | unnamed list. onExpand}

\item{onFocus}{Logical | numeric | character | named list | unnamed list. onFocus}

\item{persistent}{Logical | numeric | character | named list | unnamed list. persistent}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{searchContainerClass}{Logical | numeric | character | named list | unnamed list. searchContainerClass}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{tabIndex}{Logical | numeric | character | named list | unnamed list. tabIndex}

\item{translateWithId}{Logical | numeric | character | named list | unnamed list. translateWithId}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

