% Auto-generated: do not edit by hand
\name{tabList}

\alias{tabList}

\title{TabList component}

\description{
TabList is a wrapper for the Carbon TabList component.
}

\usage{
tabList(children=NULL, id=NULL, activation=NULL, className=NULL,
contained=NULL, fullWidth=NULL, iconSize=NULL,
leftOverflowButtonProps=NULL, light=NULL,
loading_state=NULL, rightOverflowButtonProps=NULL,
scrollDebounceWait=NULL, scrollIntoView=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{activation}{Logical | numeric | character | named list | unnamed list. activation}

\item{className}{Character. className}

\item{contained}{Logical. contained}

\item{fullWidth}{Logical | numeric | character | named list | unnamed list. fullWidth}

\item{iconSize}{A list of or a singular dash component, string or number. iconSize}

\item{leftOverflowButtonProps}{Logical | numeric | character | named list | unnamed list. leftOverflowButtonProps}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{rightOverflowButtonProps}{Logical | numeric | character | named list | unnamed list. rightOverflowButtonProps}

\item{scrollDebounceWait}{Logical | numeric | character | named list | unnamed list. scrollDebounceWait}

\item{scrollIntoView}{Logical | numeric | character | named list | unnamed list. scrollIntoView}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

