% Auto-generated: do not edit by hand
\name{tabs}

\alias{tabs}

\title{Tabs component}

\description{
Tabs is a wrapper for the Carbon Tabs component.
}

\usage{
tabs(children=NULL, id=NULL, className=NULL,
defaultSelectedIndex=NULL, dismissable=NULL,
loading_state=NULL, onChange=NULL, onTabCloseRequest=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, selectedIndex=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultSelectedIndex}{Logical | numeric | character | named list | unnamed list. defaultSelectedIndex}

\item{dismissable}{Logical | numeric | character | named list | unnamed list. dismissable}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onTabCloseRequest}{Logical | numeric | character | named list | unnamed list. onTabCloseRequest}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{selectedIndex}{Numeric. selectedIndex}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

