% Auto-generated: do not edit by hand
\name{tabsSkeleton}

\alias{tabsSkeleton}

\title{TabsSkeleton component}

\description{
TabsSkeleton is a wrapper for the Carbon TabsSkeleton component.
}

\usage{
tabsSkeleton(children=NULL, id=NULL, className=NULL, contained=NULL,
loading_state=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{contained}{Logical | numeric | character | named list | unnamed list. contained}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

