% Auto-generated: do not edit by hand
\name{tabsVertical}

\alias{tabsVertical}

\title{TabsVertical component}

\description{
TabsVertical is a wrapper for the Carbon TabsVertical component.
}

\usage{
tabsVertical(children=NULL, id=NULL, className=NULL,
defaultSelectedIndex=NULL, height=NULL, loading_state=NULL,
onChange=NULL, selectedIndex=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultSelectedIndex}{Logical | numeric | character | named list | unnamed list. defaultSelectedIndex}

\item{height}{Logical | numeric | character | named list | unnamed list. height}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{selectedIndex}{Logical | numeric | character | named list | unnamed list. selectedIndex}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

