% Auto-generated: do not edit by hand
\name{tag}

\alias{tag}

\title{Tag component}

\description{
Tag is a wrapper for the Carbon Tag component.
}

\usage{
tag(children=NULL, id=NULL, as_=NULL, className=NULL,
decorator=NULL, disabled=NULL, filter=NULL,
loading_state=NULL, onClose=NULL, renderIcon=NULL,
size=NULL, slug=NULL, style=NULL, title=NULL, type=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{disabled}{Logical. disabled}

\item{filter}{Logical | numeric | character | named list | unnamed list. filter}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClose}{Logical | numeric | character | named list | unnamed list. onClose}

\item{renderIcon}{Logical | numeric | character | named list | unnamed list. renderIcon}

\item{size}{Character. size}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{title}{Logical | numeric | character | named list | unnamed list. title}

\item{type}{Character. type}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

