% Auto-generated: do not edit by hand
\name{textArea}

\alias{textArea}

\title{TextArea component}

\description{
TextArea is a wrapper for the Carbon TextArea component.
}

\usage{
textArea(children=NULL, id=NULL, className=NULL, cols=NULL,
counterMode=NULL, debounce=NULL, decorator=NULL,
defaultValue=NULL, disabled=NULL, enableCounter=NULL,
helperText=NULL, hideLabel=NULL, invalid=NULL,
invalidText=NULL, label=NULL, labelText=NULL, light=NULL,
loading_state=NULL, maxCount=NULL, n_blur=NULL,
n_submit=NULL, onChange=NULL, onClick=NULL, onKeyDown=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, placeholder=NULL, readOnly=NULL,
rows=NULL, slug=NULL, style=NULL, value=NULL, warn=NULL,
warnText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{cols}{Logical | numeric | character | named list | unnamed list. cols}

\item{counterMode}{Logical | numeric | character | named list | unnamed list. counterMode}

\item{debounce}{Logical | numeric. debounce}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. defaultValue}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{enableCounter}{Logical | numeric | character | named list | unnamed list. enableCounter}

\item{helperText}{Logical | numeric | character | named list | unnamed list. helperText}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{label}{Character. label}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{maxCount}{Logical | numeric | character | named list | unnamed list. maxCount}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onKeyDown}{Logical | numeric | character | named list | unnamed list. onKeyDown}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{rows}{Logical | numeric | character | named list | unnamed list. rows}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}

\item{value}{Character. value}

\item{warn}{Logical | numeric | character | named list | unnamed list. warn}

\item{warnText}{Logical | numeric | character | named list | unnamed list. warnText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

