% Auto-generated: do not edit by hand
\name{theme}

\alias{theme}

\title{Theme component}

\description{
Theme is a wrapper for the Carbon Theme component.
}

\usage{
theme(children=NULL, id=NULL, as_=NULL, className=NULL,
loading_state=NULL, style=NULL, theme=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{className}{Character. className}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}

\item{theme}{Logical | numeric | character | named list | unnamed list. theme}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

