% Auto-generated: do not edit by hand
\name{tile}

\alias{tile}

\title{Tile component}

\description{
Tile is a wrapper for the Carbon Tile component.
}

\usage{
tile(children=NULL, id=NULL, className=NULL, decorator=NULL,
hasRoundedCorners=NULL, light=NULL, loading_state=NULL,
slug=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{decorator}{Logical | numeric | character | named list | unnamed list. decorator}

\item{hasRoundedCorners}{Logical | numeric | character | named list | unnamed list. hasRoundedCorners}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{slug}{Logical | numeric | character | named list | unnamed list. slug}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

