% Auto-generated: do not edit by hand
\name{tileGroup}

\alias{tileGroup}

\title{TileGroup component}

\description{
TileGroup is a wrapper for the Carbon TileGroup component.
}

\usage{
tileGroup(children=NULL, id=NULL, className=NULL,
defaultSelected=NULL, disabled=NULL, legend=NULL,
loading_state=NULL, name=NULL, onChange=NULL, required=NULL,
style=NULL, valueSelected=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultSelected}{Logical | numeric | character | named list | unnamed list. defaultSelected}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{legend}{Logical | numeric | character | named list | unnamed list. legend}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{name}{Logical | numeric | character | named list | unnamed list. name}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{required}{Logical | numeric | character | named list | unnamed list. required}

\item{style}{Named list. style}

\item{valueSelected}{Logical | numeric | character | named list | unnamed list. valueSelected}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

