% Auto-generated: do not edit by hand
\name{timePicker}

\alias{timePicker}

\title{TimePicker component}

\description{
TimePicker is a wrapper for the Carbon TimePicker component.
}

\usage{
timePicker(children=NULL, id=NULL, className=NULL, debounce=NULL,
disabled=NULL, hideLabel=NULL, invalid=NULL,
invalidText=NULL, labelText=NULL, light=NULL,
loading_state=NULL, maxLength=NULL, n_blur=NULL,
n_submit=NULL, onBlur=NULL, onChange=NULL, onClick=NULL,
pattern=NULL, persisted_props=NULL, persistence=NULL,
persistence_type=NULL, placeholder=NULL, readOnly=NULL,
size=NULL, style=NULL, type=NULL, value=NULL, warning=NULL,
warningText=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{invalid}{Logical | numeric | character | named list | unnamed list. invalid}

\item{invalidText}{Logical | numeric | character | named list | unnamed list. invalidText}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{light}{Logical | numeric | character | named list | unnamed list. light}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{maxLength}{Logical | numeric | character | named list | unnamed list. maxLength}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onBlur}{Logical | numeric | character | named list | unnamed list. onBlur}

\item{onChange}{Logical | numeric | character | named list | unnamed list. onChange}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{pattern}{Logical | numeric | character | named list | unnamed list. pattern}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{placeholder}{Logical | numeric | character | named list | unnamed list. placeholder}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{type}{Logical | numeric | character | named list | unnamed list. type}

\item{value}{Logical | numeric | character | named list | unnamed list. value}

\item{warning}{Logical | numeric | character | named list | unnamed list. warning}

\item{warningText}{Logical | numeric | character | named list | unnamed list. warningText}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

