% Auto-generated: do not edit by hand
\name{toastNotification}

\alias{toastNotification}

\title{ToastNotification component}

\description{
ToastNotification is a wrapper for the Carbon ToastNotification component.
}

\usage{
toastNotification(children=NULL, id=NULL, ariaLabel=NULL, caption=NULL,
className=NULL, hideCloseButton=NULL, kind=NULL,
loading_state=NULL, lowContrast=NULL, onClose=NULL,
onCloseButtonClick=NULL, role=NULL,
statusIconDescription=NULL, style=NULL, subtitle=NULL,
timeout=NULL, title=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{ariaLabel}{Logical | numeric | character | named list | unnamed list. ariaLabel}

\item{caption}{Logical | numeric | character | named list | unnamed list. caption}

\item{className}{Character. className}

\item{hideCloseButton}{Logical | numeric | character | named list | unnamed list. hideCloseButton}

\item{kind}{Logical | numeric | character | named list | unnamed list. kind}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{lowContrast}{Logical | numeric | character | named list | unnamed list. lowContrast}

\item{onClose}{Logical | numeric | character | named list | unnamed list. onClose}

\item{onCloseButtonClick}{Logical | numeric | character | named list | unnamed list. onCloseButtonClick}

\item{role}{Logical | numeric | character | named list | unnamed list. role}

\item{statusIconDescription}{Logical | numeric | character | named list | unnamed list. statusIconDescription}

\item{style}{Named list. style}

\item{subtitle}{Logical | numeric | character | named list | unnamed list. subtitle}

\item{timeout}{Logical | numeric | character | named list | unnamed list. timeout}

\item{title}{Logical | numeric | character | named list | unnamed list. title}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

