% Auto-generated: do not edit by hand
\name{toggle}

\alias{toggle}

\title{Toggle component}

\description{
Toggle is a wrapper for the Carbon Toggle component.
}

\usage{
toggle(children=NULL, id=NULL, className=NULL, defaultToggled=NULL,
disabled=NULL, hideLabel=NULL, label=NULL, labelA=NULL,
labelB=NULL, labelText=NULL, loading_state=NULL,
onClick=NULL, onToggle=NULL, persisted_props=NULL,
persistence=NULL, persistence_type=NULL, readOnly=NULL,
size=NULL, style=NULL, toggled=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{className}{Character. className}

\item{defaultToggled}{Logical | numeric | character | named list | unnamed list. defaultToggled}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{label}{Character. label}

\item{labelA}{Logical | numeric | character | named list | unnamed list. labelA}

\item{labelB}{Logical | numeric | character | named list | unnamed list. labelB}

\item{labelText}{Logical | numeric | character | named list | unnamed list. labelText}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onClick}{Logical | numeric | character | named list | unnamed list. onClick}

\item{onToggle}{Logical | numeric | character | named list | unnamed list. onToggle}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{readOnly}{Logical | numeric | character | named list | unnamed list. readOnly}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}

\item{toggled}{Logical. toggled}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

