% Auto-generated: do not edit by hand
\name{toggletip}

\alias{toggletip}

\title{Toggletip component}

\description{
Toggletip is a wrapper for the Carbon Toggletip component.
}

\usage{
toggletip(children=NULL, id=NULL, align=NULL,
alignmentAxisOffset=NULL, as_=NULL, autoAlign=NULL,
autoAlignBoundary=NULL, backgroundToken=NULL, border=NULL,
caret=NULL, className=NULL, defaultOpen=NULL,
dropShadow=NULL, highContrast=NULL, isTabTip=NULL,
loading_state=NULL, onRequestClose=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{alignmentAxisOffset}{Logical | numeric | character | named list | unnamed list. alignmentAxisOffset}

\item{as_}{Logical | numeric | character | named list | unnamed list. as}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{autoAlignBoundary}{Logical | numeric | character | named list | unnamed list. autoAlignBoundary}

\item{backgroundToken}{Logical | numeric | character | named list | unnamed list. backgroundToken}

\item{border}{Logical | numeric | character | named list | unnamed list. border}

\item{caret}{Logical | numeric | character | named list | unnamed list. caret}

\item{className}{Character. className}

\item{defaultOpen}{Logical | numeric | character | named list | unnamed list. defaultOpen}

\item{dropShadow}{Logical | numeric | character | named list | unnamed list. dropShadow}

\item{highContrast}{Logical | numeric | character | named list | unnamed list. highContrast}

\item{isTabTip}{Logical | numeric | character | named list | unnamed list. isTabTip}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{onRequestClose}{Logical | numeric | character | named list | unnamed list. onRequestClose}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

