% Auto-generated: do not edit by hand
\name{tooltip}

\alias{tooltip}

\title{Tooltip component}

\description{
Tooltip is a wrapper for the Carbon Tooltip component.
}

\usage{
tooltip(children=NULL, id=NULL, align=NULL, className=NULL,
closeOnActivation=NULL, defaultOpen=NULL, description=NULL,
dropShadow=NULL, enterDelayMs=NULL, highContrast=NULL,
label=NULL, leaveDelayMs=NULL, loading_state=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{align}{Character. align}

\item{className}{Character. className}

\item{closeOnActivation}{Logical | numeric | character | named list | unnamed list. closeOnActivation}

\item{defaultOpen}{Logical. defaultOpen}

\item{description}{Character. description}

\item{dropShadow}{Logical | numeric | character | named list | unnamed list. dropShadow}

\item{enterDelayMs}{Logical | numeric | character | named list | unnamed list. enterDelayMs}

\item{highContrast}{Logical | numeric | character | named list | unnamed list. highContrast}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{leaveDelayMs}{Logical | numeric | character | named list | unnamed list. leaveDelayMs}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

