% Auto-generated: do not edit by hand
\name{treeNode}

\alias{treeNode}

\title{TreeNode component}

\description{
TreeNode is a wrapper for the Carbon TreeNode component.
}

\usage{
treeNode(children=NULL, id=NULL, active=NULL, align=NULL,
autoAlign=NULL, className=NULL, debounce=NULL,
defaultIsExpanded=NULL, depth=NULL, disabled=NULL,
href=NULL, isExpanded=NULL, label=NULL, loading_state=NULL,
n_blur=NULL, n_submit=NULL, onNodeFocusEvent=NULL,
onSelect=NULL, onToggle=NULL, onTreeSelect=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, renderIcon=NULL, selected=NULL,
style=NULL, value=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{active}{Logical | numeric | character | named list | unnamed list. active}

\item{align}{Logical | numeric | character | named list | unnamed list. align}

\item{autoAlign}{Logical | numeric | character | named list | unnamed list. autoAlign}

\item{className}{Character. className}

\item{debounce}{Logical | numeric. debounce}

\item{defaultIsExpanded}{Logical | numeric | character | named list | unnamed list. defaultIsExpanded}

\item{depth}{Logical | numeric | character | named list | unnamed list. depth}

\item{disabled}{Logical | numeric | character | named list | unnamed list. disabled}

\item{href}{Logical | numeric | character | named list | unnamed list. href}

\item{isExpanded}{Logical | numeric | character | named list | unnamed list. isExpanded}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{n_blur}{Numeric. n_blur}

\item{n_submit}{Numeric. n_submit}

\item{onNodeFocusEvent}{Logical | numeric | character | named list | unnamed list. onNodeFocusEvent}

\item{onSelect}{Logical | numeric | character | named list | unnamed list. onSelect}

\item{onToggle}{Logical | numeric | character | named list | unnamed list. onToggle}

\item{onTreeSelect}{Logical | numeric | character | named list | unnamed list. onTreeSelect}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{renderIcon}{A list of or a singular dash component, string or number. renderIcon}

\item{selected}{Logical | numeric | character | named list | unnamed list. selected}

\item{style}{Named list. style}

\item{value}{Logical | numeric | character | named list | unnamed list. value}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

