% Auto-generated: do not edit by hand
\name{treeView}

\alias{treeView}

\title{TreeView component}

\description{
TreeView is a wrapper for the Carbon TreeView component.
}

\usage{
treeView(children=NULL, id=NULL, active=NULL, className=NULL,
hideLabel=NULL, label=NULL, loading_state=NULL,
multiselect=NULL, onActivate=NULL, onSelect=NULL,
persisted_props=NULL, persistence=NULL,
persistence_type=NULL, selected=NULL, size=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. children}

\item{id}{Character. id}

\item{active}{Logical | numeric | character | named list | unnamed list. active}

\item{className}{Character. className}

\item{hideLabel}{Logical | numeric | character | named list | unnamed list. hideLabel}

\item{label}{Logical | numeric | character | named list | unnamed list. label}

\item{loading_state}{Lists containing elements 'is_loading', 'prop_name', 'component_name'.
those elements have the following types:
  - is_loading (logical; optional)
  - prop_name (character; optional)
  - component_name (character; optional). loading_state}

\item{multiselect}{Logical | numeric | character | named list | unnamed list. multiselect}

\item{onActivate}{Logical | numeric | character | named list | unnamed list. onActivate}

\item{onSelect}{Logical | numeric | character | named list | unnamed list. onSelect}

\item{persisted_props}{List of characters. persisted_props}

\item{persistence}{Logical | character | numeric. persistence}

\item{persistence_type}{A value equal to: 'local', 'session', 'memory'. persistence_type}

\item{selected}{Logical | numeric | character | named list | unnamed list. selected}

\item{size}{Logical | numeric | character | named list | unnamed list. size}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

